% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{abundantClustersViewer}
\alias{abundantClustersViewer}
\title{abundantClustersViewer}
\usage{
abundantClustersViewer(AC, show.cluster.sizes = TRUE,
  show.all.labels = FALSE)
}
\arguments{
\item{AC}{an object of class AC (object returned by the 'computeAC()' function)}

\item{show.cluster.sizes}{a logicial specifing if points size are related to cell count or not (TRUE by default)}

\item{show.all.labels}{a logicial specifing if all cluster label must be show or just significant cluster}
}
\value{
a ggplot object
}
\description{
Generate a scatter plot representation displaying for each cluster its m
}
\details{
By default, only significant abundant clusters are labeled. Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}

