% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{AC-class}
\alias{AC}
\alias{AC-class}
\title{Abundant Clusters (AC) class definition}
\description{
The Results object is a S4 object containing the results identification of abundant cluster. 
It also contains all informations about statistical parameters used to performs this test.
}
\details{
AC is a printable and a plotable object calling the abundantClustersViewer() fonction.
}
\section{Slots}{

\describe{
\item{\code{sample.names}}{a character vector containing the samples names used}

\item{\code{cluster.size}}{a numeric vector containing number of cells ( sum of all samples ) for each cluster}

\item{\code{use.percentages}}{a logical specifying if computations was performed on percentage of cell abondance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the AC}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used}

\item{\code{th.pvalue}}{a numeric vector with pvalue threshold}

\item{\code{th.mean}}{a numeric vector with mean threshold}

\item{\code{result}}{a data.frame with clusters in row. The first and second colunm contain the mean and the standard deviation and the third contain the pvalue. Finnaly the last columns is a logicial ("significance") describing if the 2 thresholds was reached or not}
}}

