% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{annotateClusters}
\alias{annotateClusters}
\title{Annotating cell clusters}
\usage{
annotateClusters(Results, annotations, num = 5)
}
\arguments{
\item{Results}{a Results object}

\item{annotations}{a character dataframe specifying the annotations}

\item{num}{a numeric value specifying the number of markers expression categories to be used}
}
\value{
a Results object
}
\description{
This function is used to annotate cell clusters based on their marker expression categories.
The annotations of cell clusters must be specifyed using a character dataframe indicating for each marker of each population the matching categorial values.
}
\details{
The annotation data.frame must have the annotations in rows and the markers in colomuns.
Matching values must be specified as a character list of acceptable values.
}

