% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{AP-class}
\alias{AP}
\alias{AP-class}
\title{Abundance Profiles (AP) class definition}
\description{
The 'AP' object is a S4 object containing the information related to the cluster classification based on theirs abundance profiles.
Moreover, this object contains parameters and results used in the statistical analysis.
}
\details{
Five methods are available to classify cellular clusters: 'hierarchical_k', 'hierarchical_h', 'kmeans', 'eigencell' and 'clique'.
Each method can parameterized using the 'method.parameter' parameter.
 
The 'print()' and 'show()' can be used to display a summary of this object. 
Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'classifyClusteringResults()' function.
}
\section{Slots}{

\describe{
\item{\code{class.number}}{a numeric value specifying the number of clusters}

\item{\code{cluster.size}}{a numeric vector containing the number of cells for each cluster}

\item{\code{method}}{a character specifying the method used to classify cluster}

\item{\code{method.parameter}}{a named list of parameters used by the classification method}

\item{\code{classes}}{a two column dataframe with the cluster in first column and corresponding class in the second column}
}}

