% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{assignContext}
\alias{assignContext}
\alias{assignContext,Results-method}
\title{Assignement of a context into a 'Results' object}
\usage{
assignContext(Results, assignments)

\S4method{assignContext}{Results}(Results, assignments)
}
\arguments{
\item{Results}{a Results or Results object}

\item{assignments}{a data.frame containing all samples names (in rownames) and columns providing contextual associations like "bc" (biological conditions), "tp" (biological conditions) and "ind" (individuals)}
}
\value{
none
}
\description{
This function allows to assigns three kinds of contextual information to samples: biological condition, timepoints and individuals.
}
\details{
The order of rownames (sample names) in this assignments dataframe will be the display order of the elements.
}

