% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildCircles}
\alias{buildCircles}
\title{Internal - Generate a circle representation}
\usage{
buildCircles(circles, color = "grey80", class = NA, npoint = 100,
  limits = 30000, maxiter = 100)
}
\arguments{
\item{circles}{a 2 column dataframe the clusters to be displayed and theirs sizes}

\item{color}{a character specifying the color of the packed circles representation}

\item{class}{a numeric specifying the class number to be displayed}

\item{npoint}{a numeric specifying the levels of details of polygons}

\item{limits}{a numeric specifying the size of the coordinate system centered on (0,0)}

\item{maxiter}{a numeric specifying the maximal number of iterations to perform}
}
\value{
a ggplot2 object
}
\description{
This function is used internally to generate a packed circles representation
}

