% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildCirclesLegend}
\alias{buildCirclesLegend}
\title{Internal - Generate an legend for circles representation}
\usage{
buildCirclesLegend(circles = data.frame(x = c(-29500, -19000, -8000, 3000,
  20000), y = c(20000, 20000, 20000, 20000, 20000), r = c(500, 1000, 2000, 5000,
  10000)), npoint = 100, limits = 30000)
}
\arguments{
\item{circles}{a 3 columns data frame with the x, y coordinate of points and their radius}

\item{npoint}{a numeric specifying the levels of details of polygons}

\item{limits}{a numeric specifying the size of the coordinate system centered on (0,0)}
}
\value{
a ggplot2 object
}
\description{
This function is used internally to generate the legend of a packed circles representation
}

