% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{CC-class}
\alias{CC}
\alias{CC-class}
\title{Correlated Clusters (CC) class definition}
\description{
The CC object is a S4 object containing the identified correlated clusters with a numeric vector of phenotypic data. 
It contains all informations about statistical parameters used to performs this test.
}
\details{
CC is a printable and a plotable object calling the correlatedClustersViewer() fonction
}
\section{Slots}{

\describe{
\item{\code{cluster.size}}{a numeric vector containing number of cells ( sum of all samples ) for each cluster}

\item{\code{variable}}{a numeric vector containing the expression values of the variable}

\item{\code{use.percentages}}{a logical specifying if computations was performed on percentage of cell abondance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the CC}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used}

\item{\code{th.pvalue}}{a numeric vector with pvalue threshold}

\item{\code{th.correlation}}{a numeric vector with correlation threshold (r)}

\item{\code{result}}{a three colmuns dataframe with clusters in row. The first colunm contains the coefficiant of correlation (r), the second contains the associated pvalue and the third a logical (significance) specifying if the two thresholds was reached or not.}
}}

