% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{circlesPackingViewer}
\alias{circlesPackingViewer}
\title{Visualization of abundance profiles classification}
\usage{
circlesPackingViewer(AP, show.cluster_sizes = TRUE, show.on_device = TRUE)
}
\arguments{
\item{AP}{an object of class 'AP' (object returned by the 'classifyAbundanceProfiles()' function)}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to number of associated cells}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generates a colored circle packing representation showing for each cluster: its associated class based on its abundance profile.
}
\details{
The sizes of the dots are proportionals to the total number of associated cells to each cluster.
Circle packing classes are sorted by the number of clusters in each class.
}

