% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{classificationViewer}
\alias{classificationViewer}
\title{classificationViewer}
\usage{
classificationViewer(CCR, show.on_device = TRUE)
}
\arguments{
\item{CCR}{an object of class 'CCR' (object returned by the 'classifyclusteringResults()' function)}

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generate a graph representation of classified clusters
}
\details{
Clusters of the same classe are shown using a circular graph. Circular graphs are sorted by the number of cluster in each class.
}

