% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{classifyAbundanceProfiles}
\alias{classifyAbundanceProfiles}
\title{Classification of clustering results based on the abundance profiles}
\usage{
classifyAbundanceProfiles(Results, method = "hierarchical_h",
  method.parameter = NULL)
}
\arguments{
\item{Results}{a 'Results' object}

\item{method}{a character specifying the clustering method among: "hierarchical_h", "hierarchical_k","k-means","eigencell","clique"}

\item{method.parameter}{a numeric specifying the numeric value required by the selected method}
}
\value{
a S4 object of class 'AP'
}
\description{
Classifies clusters based on their abundance profiles (number of cells for each cluster).
}
\details{
The classification is done on cell abundances of each clusters and could be performed using 5 methods:
\itemize{
\item "hierarchical_k" 
This method first compute the Pearson correlation matrix and then use this matrix to performs a hierarchical classification. 
The hierarchical classification is cut in order to return the desired number of classes. 
This number of classes must be provided as a numeric integer using the 'method.parameter' parameter.
It is to note that negative correlations are considered as uncorrelated
\item "hierarchical_h" (default method)
This method works in the same way than 'hierarchical_k' but the height where the hierarchical tree is specified. 
This height is a correlation threshold (a numeric double between 0 and 1 included, default is 0.7) provided using the 'method.parameter' parameter.
\item "kmeans"
This method works as described in the R stats documentation (?kmeans) using the 'method.parameter' parameter to specify the desired number of classes.
\item "eigencell" 
This method performs an eigen vector decomposition and then calculate the correlations between cluster values and these vectors.
Clusters which correlate above a specific threshold with the same eigen vector are classified together.
This correlation threshold (a numeric double between 0 and 1 included, default is 0.8) provided using the 'method.parameter' parameter.
\item "clique" 
This method first computes the Pearson correlation matrix and then use this matrix to generate an undirected graph.
In this graph, an edge is drawn between two nodes if the correlation coefficient in the adjacency matrix is above a specific threshold. 
This correlation threshold (a numeric double between 0 and 1 included, default is 0.7) provided using the 'method.parameter' parameter.
After building the graph, the method looking for the largest cliques which are considered as classes of nodes. Cliques correspond to subgraph in which every two distinct vertices are adjacent.
}
}
