% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{classifyEnrichmentProfiles}
\alias{classifyEnrichmentProfiles}
\title{classifyEnrichmentProfiles xxx}
\usage{
classifyEnrichmentProfiles(Results, method = "hierarchical_h",
  class.number = NULL, eigencell.correlation.th = 0.8,
  clique.correlation.th = 0.7, hierarchical.correlation.th = 0.8)
}
\arguments{
\item{Results}{a Results or SPADEResults object}

\item{method}{a character specifying the clustering method among one of those : 'hierarchical','kmeans','eigencell','clique'}

\item{class.number}{a numeric specifying the number of classes needed when the method parameter choosen is either 'hierarchical_k' or 'kmeans'}

\item{eigencell.correlation.th}{a numeric (ignored if method is not 'eigencell') specifying the correlation threshold (in [0,1], 0.8 by default)  in case of eigencell clustering}

\item{clique.correlation.th}{a numeric (ignored if method is not 'clique') specifying the correlation threshold (in [0,1], 0.7 by default) in case of clique clustering}

\item{hierarchical.correlation.th}{a numeric (ignored if method is not 'hierarchical_h') in [0,1]) specifying the threshold of correlation (in [0,1], 0.8 by default) use to cut the hirerchical tree}
}
\value{
a EnrichmentProfiles object
}
\description{
xxx
}
\details{
xxx
}

