% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{classifyPhenoProfiles}
\alias{classifyPhenoProfiles}
\title{Classification of clusters based on their phenotype profiles}
\usage{
classifyPhenoProfiles(Results, method = "hierarchical_h",
  method.parameter = NULL)
}
\arguments{
\item{Results}{a Results or SPADEResults object}

\item{method}{a character specifying the clustering method among one of those : "hierarchical_h", "hierarchical_k","k-means","eigencell","clique"}

\item{method.parameter}{a numeric specifying the numeric value required by the selected method}
}
\value{
a S4 object of class 'PhenoProfiles'
}
\description{
Classifies clusters based on their phenotype profiles.
}
\details{
The classification is done on cell abundances of each clusters and could be performed using 5 methods:
\itemize{
\item "hierarchical_k" 
This method first compute the pearson correlation matrix and then use this matrix to performs a hierarchical classification. 
The hierarchical classification is cutted is order provided the desired number of classes. 
This number of classes must be provided as a numeric integer using the 'method.parameter' parameter.
It is to note that negative correlations are considered as uncorrelated;
\item "hierarchical_k" 
This method works in the same way than 'hierarchical_k' but the height where the hierarchical tree is specified. 
This heigth is correlation threshold and must be a numeric double (between 0 and 1 included) provided using the 'method.parameter' parameter.
\item "eigencell" 
This method compute the performs a eigen vector decomposition and then calculate the correlations between cluster expressions and these vectors.
Clusters which correlate above a specific threshold with the same eigen vector are classified together.
This correlation threshold must be a numeric double (between 0 and 1 included) provided using the 'method.parameter' parameter.
\item "clique" 
This method first compute the pearson correlation matrix and then use this matrix to generate an undirected graph.
In this graph, an edge is drawn between two nodes if the correlation coefficient in the adjacency matrix is above a specific threshold. 
This correlation threshold must be a numeric double (between 0 and 1 included) provided using the 'method.parameter' parameter.
The largest cliques are then researched in this graph. Cliques correspond to subgraph in which every two distinct vertices are adjacent.
Finally the founded cliques are considered as classes of cluster.
}
}

