% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_viewers.R
\name{clusterViewer}
\alias{clusterViewer}
\title{Visualization of cluster phenotypes}
\usage{
clusterViewer(Results, clusters = NULL, samples = NULL, markers = NULL,
  show.mean = "both")
}
\arguments{
\item{Results}{a SPADEResults or Result object}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{samples}{named vector providing the correspondence between samples name (in rowname) and the logical value TRUE to use these samples (all samples by default)}

\item{markers}{a character vector specifying the markers to be displayed}

\item{show.mean}{a character specifying if marker means expression should be displayed, possible value are among : "none", "only" or "both"}
}
\value{
a list of 'ggplot' objects
}
\description{
Generates a parallel coordinate plot representation showing for each cluster the marker median expressions.
}
\details{
If 'Results' is an object of class 'SPADEResults' then the quantile ranges will be displayed using a ribbon.

The 'samples' parameter required a named vector providing the correspondence between a sample name (in rowname) and the logical value TRUE to show this sample or FALSE otherwise.

The 'show.mean' parameter allows to visualize three kinds of information:
\itemize{
\item "none" value will show marker median expressions for each selected samples;
\item "only" value will show only the mean of median maker expressions for all selected samples (displayed as black dashed line);
\item "both" value will show marker median expressions for each selected samples together with the mean of median maker expressions for all selected samples.
}
}

