% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeAC}
\alias{computeAC}
\title{Compute the Abundant Clusters}
\usage{
computeAC(Results, condition, use.percentages = TRUE, method = "t.test",
  method.adjust = NULL, th.pvalue = 0.05, th.mean = 0)
}
\arguments{
\item{Results}{a Results or SPADEResults object}

\item{condition}{a named vector providing the correspondence between a sample name (in rowname) and the logical value TRUE to test abondance for this sample or FALSE otherwise}

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character containing the statistical method to use for the ACs detection. The parameter can take the values "t.test" or "wilcox.test"}

\item{method.adjust}{a character specifying if the pvalues should be corrected using the argument "method" available for function p.adjust, among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr"}

\item{th.pvalue}{a numeric specifying the pvalue threshold (0.05 by default)}

\item{th.mean}{a numeric specifying the mean threshold (0 by default)}
}
\value{
a AC object
}
\description{
Abundant Clusters are clusters wich have an enrichement significatyvely different of zeor
}
\details{
xxx
}

