% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeCC}
\alias{computeCC}
\title{Compute the correlation of SPADE cluster with a cynetics phenotype}
\usage{
computeCC(Results, variable, use.percentages = TRUE, method = "pearson",
  method.adjust = NULL, th.pvalue = 0.05, th.correlation = 0.5)
}
\arguments{
\item{Results}{a Results or SPADEResults object}

\item{variable}{a numerical named vector providing the correspondence between a sample name (in rowname) and the specific phenotype or NA to ignore a sample}

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character indicating the correlation method to use : "pearson", "spearman"}

\item{method.adjust}{a character specifying if the pvalues should be corrected using the argument "method" available for function p.adjust, among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr"}

\item{th.pvalue}{a numeric specifying the pvalue threshold (0.05 by default)}

\item{th.correlation}{a numeric specifying the r two sided threshold (0.5 by default) in [0,1]}
}
\value{
a CC object
}
\description{
Correlated Clusters are clusters with count or % correlated with a specific phenotipical variable (such as titer)
}
\details{
xxx
}

