% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeEigenCellClusters}
\alias{computeEigenCellClusters}
\title{Internal - computeEigenCellClusters}
\usage{
computeEigenCellClusters(data, eigencell.correlation.th = 0.8)
}
\arguments{
\item{data}{a matrix with all clusters in rownames}

\item{eigencell.correlation.th}{}
}
\value{
a dataframe containing the cluster ID, the classe of this cluster and a clustering score.
}
\description{
xxx
}
\details{
xxx
}

