% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeKmeans}
\alias{computeKmeans}
\title{Internal - Kmeans classification}
\usage{
computeKmeans(data, k = NULL)
}
\arguments{
\item{data}{a numeric matrix with cluster names in rownames}

\item{k}{a numeric specifying the desired number of classes}
}
\value{
a dataframe containing for each cluster, its name and class
}
\description{
This function is used internally to classify clusters abundance profiles or phenotype profiles using a k-means algorithm.
}
\details{
This method works as described in the R stats documentation (?kmeans) using the 'k' parameter to specify the desired number of classes.
}

