% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{computeNumberofClustersHavingLessThatThreshold}
\alias{computeNumberofClustersHavingLessThatThreshold}
\title{Computation of the fraction of cluster with low number of cells}
\usage{
computeNumberofClustersHavingLessThatThreshold(SPADEResults, th = 5)
}
\arguments{
\item{SPADEResults}{a SPADEResults object}

\item{num}{a numeric value specifying the cell threshold}
}
\value{
a numeric value of the fraction of clusters having less cells that the specific threshold
}
\description{
Computes the fraction of clusters having a number of cells less than a specific threshold
}
\details{
xxx
}

