% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{computeNumberofClustersHavingMultiPop}
\alias{computeNumberofClustersHavingMultiPop}
\title{Computation of the fraction of cluster with multiple population}
\usage{
computeNumberofClustersHavingMultiPop(SPADEResults, class = 5)
}
\arguments{
\item{SPADEResults}{a SPADEResults object}

\item{class}{a numeric value specifying the number of classes to found}
}
\value{
a numeric value of the fraction of clusters having less cells that the specific threshold
}
\description{
Computes the fraction of clusters having a number of cells less than a specific threshold
}
\details{
xxx
}

