% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{computePhenoTable}
\alias{computePhenoTable}
\title{Internal - computePhenoTable}
\usage{
computePhenoTable(SPADEResults, num = 5)
}
\arguments{
\item{SPADEResults}{a SPADEResults object}

\item{num}{a numeric indicating the number of "phenotype"}
}
\value{
xxx
}
\description{
computePhenoTable
}
\details{
xxx
}

