% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{computeQuantile.approximation}
\alias{computeQuantile.approximation}
\title{Internal - Compute quantile with FCS flowset sample by sample}
\usage{
computeQuantile.approximation(flowset, probs = c(0.05, 0.95))
}
\arguments{
\item{flowset}{a flowCore flowset}

\item{probs}{a numeric vector of 2 values specifying the quantiles to compute}
}
\value{
a numeric matrix containing the quantiles of each marker
}
\description{
This function is used internally to provide the mean of quantiles from each sample to seed up computation.
}
\details{
This function performs an approximate calculation of quantiles using less memory than computeQuantile.
}

