% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{computeQuantile.heuristic}
\alias{computeQuantile.heuristic}
\title{Internal - Compute quantile with FCS flowset sample by sample}
\usage{
computeQuantile.heuristic(flowset, probs = c(0.05, 0.95))
}
\arguments{
\item{flowset}{a Flowcore flowset}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute quantiles}
}
\value{
a numeric matrix with bounds
}
\description{
This function is used internally, it provide the mean of quantiles from each sample to seed up computation
}

