% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{computeQuantile}
\alias{computeQuantile}
\title{Internal - Compute quantile with FCS flowset marker by marker}
\usage{
computeQuantile(flowset, probs = c(0.05, 0.95))
}
\arguments{
\item{flowset}{a flowCore flowset}

\item{probs}{a numeric vector of 2 values specifying the quantiles to compute}
}
\value{
a numeric matrix containing the quantiles of each marker
}
\description{
This function is used internally to compute the maker range quantiles.
}
\details{
This function performs the exact calculation of quantiles with all cells but needs more resources (time and memory usage) than 'computeQuantile.approximation'.
}

