% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{correlatedClustersViewer}
\alias{correlatedClustersViewer}
\title{correlatedClustersViewer}
\usage{
correlatedClustersViewer(CC, cluster.size = TRUE, all.label = FALSE)
}
\arguments{
\item{CC}{the CC object}

\item{cluster.size}{a logicial specifing if points size are related to cell count or not (TRUE by default)}

\item{all.label}{a logicial specifing if all cluster label must be show or just significant cluster}
}
\value{
a ggplot object
}
\description{
Generate a representation to compare cluster abondances
}
\details{
xxx
}

