% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DAC-class}
\alias{DAC}
\alias{DAC-class}
\title{Differentially Abundant Clusters (DAC) class definition}
\description{
The 'DAC' object is a S4 object containing the information related to the differentially abundant clusters between two given biological conditions. 
Moreover, this object contains parameters and results used in the statistical analysis.
}
\details{
A cluster is considered as a differentially enriched cluster if its associated p-value and fold-change are below the specific thresholds 'th.pvalue' and 'th.fc'.  

The 'print()' and 'show()' can be used to display a summary of this object.
Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'identifyDAC()' function.
}
\section{Slots}{

\describe{
\item{\code{sample.cond1}}{a character specifying the names of the samples of the first biological condition}

\item{\code{sample.cond2}}{a character specifying the names of the samples of the second biological condition}

\item{\code{cluster.size}}{a numeric vector containing the number of cells for each cluster}

\item{\code{use.percentages}}{a logical specifying if computation was performed on percentage of cell abundance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the DAC}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used (if any)}

\item{\code{method.paired}}{a logical indicating if the statistical test have been performed in a paired manner}

\item{\code{th.fc}}{a numeric value specifying the fold-change threshold}

\item{\code{th.pvalue}}{a numeric value specifying the p-value threshold}

\item{\code{results}}{a data.frame containing for each cluster (first column): the fold-change (second column) and the standard deviation (third column) for the first biological condition, the fold-change (fourth column) and the standard deviation (fifth column) for the second biological condition, the associated p-value (sixth column) and a logical (seventh column) specifying if the cluster is significantly differentially abundant.}
}}

