% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{dendro}
\alias{dendro}
\title{Internal - Build a dendrograms plot}
\usage{
dendro(ddata, row = !col, col = !row)
}
\arguments{
\item{ddata}{xxx}

\item{row}{xxx}

\item{col}{xxx}
}
\value{
a dendroplot
}
\description{
xxx
}
\details{
xxx
}

