% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{distogramViewer}
\alias{distogramViewer}
\title{Visualization of marker co-expressions}
\usage{
distogramViewer(Results, clusters = NULL, samples = NULL)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{clusters}{xxx}

\item{samples}{xxx}
}
\value{
a list of 'ggplot' objects
}
\description{
Generate a distogram representation showing the marker co-expression.
}
\details{
xxx
}

