% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{EnrichmentProfiles-class}
\alias{EnrichmentProfiles}
\alias{EnrichmentProfiles-class}
\title{EnrichmentProfiles class definition}
\description{
EnrichmentProfiles is a S4 object containing the result of classifyEnrichmentProfiles() function.
It classify clusters by their cells enrichment properties across the samples.
}
\details{
EnrichmentProfiles is a printable and a plotable object calling the profileViewer() fonction
}
\section{Slots}{

\describe{
\item{\code{method}}{a character specifying the method used to classify cluster}

\item{\code{method.parameter}}{a list of parameters used by the selected method}

\item{\code{cluster.size}}{a numeric vector with the number of cell in each cluster (sum of all samples)}

\item{\code{cluster.number}}{a numeric providing the number of cluster}

\item{\code{class.number}}{a numeric providing the number of classes}

\item{\code{classes}}{a two column dataframe with the cluster in first colunm and corresponding classe in the second colunm}
}}

