% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{exclude.markers}
\alias{exclude.markers}
\title{Internal - Removing of cell markers to exclude from a matrix}
\usage{
exclude.markers(data, exclude, colnames.FCS = NULL)
}
\arguments{
\item{data}{a numeric matrix or flowset}

\item{exclude}{a character vector containing the cell markers to be excluded}

\item{colnames.FCS}{a character vector containing colnames if data is a FCS flowset}
}
\value{
a numeric matrix without the cell markers to exclude
}
\description{
This function is used internally to remove one or several cell markers.
}
\details{
xxx
}

