% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,AC-method}
\alias{export,CC-method}
\alias{export,DEC-method}
\alias{export,EnrichmentProfiles-method}
\alias{export,PhenoProfiles-method}
\alias{export,Results-method}
\title{Exportation of SPADEVizR objects}
\usage{
export(object, filename)

\S4method{export}{Results}(object, filename = "Results.txt")

\S4method{export}{AC}(object, filename = "AC.txt")

\S4method{export}{DEC}(object, filename = "DEC.txt")

\S4method{export}{CC}(object, filename = "CC.txt")

\S4method{export}{PhenoProfiles}(object, filename = "PhenoProfiles.txt")

\S4method{export}{EnrichmentProfiles}(object,
  filename = "EnrichmentProfiles.txt")
}
\arguments{
\item{object}{a SPADEVizR object}

\item{filename}{a character indicating the location of output file}
}
\value{
none
}
\description{
Exports an object into a TAB delimeted file with parameters in header (begining with a sharp '#')
}

