% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{filter.medians}
\alias{filter.medians}
\title{Internal - filter medians to exclude from a matrix}
\usage{
filter.medians(data, use.raw.medians = FALSE)
}
\arguments{
\item{data}{a SPADE matrix}

\item{use.raw.medians}{a logicial specifying if "transformed" or "raw" medians will be use (FALSE by default)}
}
\value{
a numeric matrix without the cell markers to exclude
}
\description{
This function is used internally to remove raw or transform medians from SPADE matrix. CVS medians are always removed.
}

