% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_axis}
\alias{g_axis}
\title{Internal - Extraction of ggplot axes}
\usage{
g_axis(gplot, x = !y, y = !x)
}
\arguments{
\item{x}{a logical, TRUE to extract x axis}

\item{y}{a logical, TRUE to extract y axis}

\item{a.gplot}{a ggplot2 plot}
}
\value{
an object of class xxx
}
\description{
This function is used internally to extract axes from a 'ggplot' objet.
}
\details{
It is to note that x and y are mutuality excluded (both cannot be both TRUE) with priority to x.
}

