% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_dendro}
\alias{g_dendro}
\title{Internal - Build a dendrograms plot}
\usage{
g_dendro(dist, row = !col, col = !row)
}
\arguments{
\item{dist}{a numeric matrix containing distances between objects}

\item{row}{a logical value specifying if the horizontal dendrogram must be computed}

\item{col}{a logical value specifying if the vertical dendrogram must be computed}
}
\value{
a 'ggplot' dendrogram object
}
\description{
This function is used internally to generate a 'ggplot' dendrogram.
}
\details{
It is to note that 'row' and 'col' are mutuality excluded (both cannot be both TRUE) with priority to row.
}

