% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_legend}
\alias{g_legend}
\title{title Internal - g_legend}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{a ggplot2 plot}
}
\value{
a ggplot2 legend
}
\description{
Extract legend from a plot
}
\details{
xxx
}

