% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{ggheatmap.plot}
\alias{ggheatmap.plot}
\title{title Internal - ggheatmap.plot}
\usage{
ggheatmap.plot(L, col.width = 0.15, row.width = 0.15)
}
\arguments{
\item{L}{the list of ggplot object provided by ggheatmap}

\item{col.width}{size of horizontal dendrogram}

\item{row.width}{size of vertical dendrogram}
}
\value{
a ggplot2 axis
}
\description{
ggheatmap.plot display the heatmap build by ggheatmap
}
\details{
xxx
}

