% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Internal - Create a list of elements allowing to build a heatmap}
\usage{
ggheatmap(matrix, dendrogram.type = "rectangle", num = 5,
  clustering.markers = NULL)
}
\arguments{
\item{matrix}{a numeric matrix containing the markers expression categories}

\item{dendrogram.type}{a character specifying the look of dendrograms ("rectangle" or "triangle", "rectangle" by default)}

\item{num}{a numeric value specifying the number of markers expression categories}

\item{clustering.markers}{a character vector of clustering markers}
}
\value{
a list of 3 plots (top dendrogram, right dendrogram, heatmap)
}
\description{
This function is used internally to build the element needed for an heatmap
}

