% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{heatmapViewer}
\alias{heatmapViewer}
\title{Visualization of all clusters phenotypes as an heatmap}
\usage{
heatmapViewer(Results, num = 5, show.on_device = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{num}{a numeric value specifying the number of markers expression categories to use}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a list of 'ggplot' objects
}
\description{
Generates an heatmap representation showing for all clusters the marker median expressions.
}
\details{
For each marker, median expressions are discretized in severals categories corresponding to the heat intensities. 
This number of categories is provided using 'num' parameter.
 
If the 'Results' parameter is a 'SPADEResults' object, markers used by SPADE to clustered cell populations are shown in bold.
}

