% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{identifyAC}
\alias{identifyAC}
\title{Identification of the Abundant Clusters}
\usage{
identifyAC(Results, samples, use.percentages = TRUE, method = "t.test",
  method.adjust = NULL, th.pvalue = 0.05, th.mean = 0)
}
\arguments{
\item{Results}{a 'Results' or 'SPADEResults' object}

<<<<<<< HEAD
\item{samples}{a named vector providing the correspondence between a sample name (in row names) and the logical value TRUE to test abundance for this sample}
=======
\item{samples}{a named vector providing the correspondence between a sample name (in row names) and the logical value TRUE to test abundance for this sample or FALSE otherwise}
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character specifying the statistical method used to identify the abundant clusters. The parameter can take the values "t.test" or "wilcox.test"}

\item{method.adjust}{a character specifying if the p-values should be corrected using multiple correction methods among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" and "fdr" (from 'stats::p.adjust' method)}

\item{th.pvalue}{a numeric specifying the p-value threshold}

\item{th.mean}{a numeric specifying the abundance mean threshold}
}
\value{
a S4 object of class 'AC'
}
\description{
This function is used to identify the abundant clusters. That is to say clusters that have cell abundance statistically greater than a specific threshold.
}

