% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{identifyCC}
\alias{identifyCC}
\title{Identification of the correlation of SPADE cluster with a phenotype}
\usage{
identifyCC(Results, variable, use.percentages = TRUE, method = "pearson",
  method.adjust = NULL, th.pvalue = 0.05, th.correlation = 0.75)
}
\arguments{
\item{Results}{a 'Results' or 'SPADEResults' object}

<<<<<<< HEAD
\item{variable}{a numerical named vector providing the correspondence between a sample name (in rownames) and the specific numerical phenotype}
=======
\item{variable}{a numerical named vector providing the correspondence between a sample name (in rownames) and the specific phenotype or NA to ignore a sample}
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character indicating the correlation method to use : "pearson", "spearman"}

\item{method.adjust}{a character specifying if the p-values should be corrected using multiple correction methods among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" and "fdr" (from 'stats::p.adjust' method)}

\item{th.pvalue}{a numeric specifying the p-value threshold}

\item{th.correlation}{a numeric specifying the absolute value of the correlation coefficient threshold}
}
\value{
a S4 object of class 'CC'
}
\description{
This function is used to identify correlated clusters. 
That is to say clusters that correlate with a phenotypic variable.
}

