% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{identifyDAC}
\alias{identifyDAC}
\title{Identification of the Differentially Abundant Clusters}
\usage{
identifyDAC(Results, condition1, condition2, use.percentages = TRUE,
  method = "t.test", method.adjust = NULL, method.paired = FALSE,
  th.pvalue = 0.05, th.fc = 1)
}
\arguments{
\item{Results}{a 'Results' or 'SPADEResults' object}

\item{condition1}{a character vector providing the sample names defined as the first condition}

\item{condition2}{a character vector providing the sample names defined as the second condition}

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character specifying the name of the statistical test to use "t.test" or "wilcox.test"}

\item{method.adjust}{a character specifying if the p-values should be corrected using multiple correction methods among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" and "fdr" (from 'stats::p.adjust' method)}

\item{method.paired}{a logical indicating if the statistical test must be performed in a paired manner}

\item{th.pvalue}{a numeric specifying the p-value threshold}

\item{th.fc}{a numeric specifying the fold-change threshold}
}
\value{
a S4 object of class 'DAC'
}
\description{
This function is used to identify differentially abundant clusters. 
That is to say clusters that are differentially abundant between two biologicals conditions.
}

