% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{importResults}
\alias{importResults}
\title{importSPADEResults}
\usage{
importResults(marker.expressions, cells.count)
}
\arguments{
\item{marker.expressions}{a numerical dataframe containing median expression values for each marker of each sample, in additions of markers, the 2 two first columns are "cluster" and "sample"}

\item{cells.count}{a matrix of cells abondances with clusters in row and samples in column}
}
\value{
Results a result object
}
\description{
Import the SPADE results from a specified path to a SPADEResult object.

Generate a SPADEresults object based on SPADE results.

The tables are loaded, the SPADE tree is loaded as well as the layout
}
\details{
We advice to use quantile.heuristic = TRUE because it is faster and much more efficient in term of memory with a low impact on precision.
}

