% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importResultsFromCLR.ACS}
\alias{importResultsFromCLR.ACS}
\title{Importation of clustering results from a folder containing ACS files}
\usage{
importResultsFromCLR.ACS(path, prob.th = 0.8, th.min_cells = 0,
  probs = c(0.05, 0.95), use.raw.medians = TRUE,
  quantile.approximation = FALSE, exclude.markers = NULL,
  assignments = NULL)
}
\arguments{
\item{path}{a character specify the path of the ACS CLR files}

\item{prob.th}{a numeric specifying the probability threshold to use for considering considering the cluster of each each cells}

\item{th.min_cells}{a numeric specifying the minimum number of cell in a cluster of a sample to take its phenotype in account}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute marker range quantiles. First is the lower bound and second is the upper bound.}

\item{use.raw.medians}{a logical specifying if arcsinh transformed or raw medians will be used in the cluster expression matrix (FALSE by default)}

\item{quantile.approximation}{a logical specifying if marker range quantiles are computed using all cells (FALSE), or is the means of the quantile of each samples (TRUE)}

\item{exclude.markers}{a character vector of markers to exclude (case insensitive)}

\item{assignments}{a data.frame containing all samples names (in rownames) and columns providing contextual associations like "bc" (biological conditions), "tp" (biological conditions) and "ind" (individuals)}
}
\value{
a S4 object of class 'Results'
}
\description{
The 'importResultsFromCLR.ACS()' function imports cell clustering results from CLR files contained in a specified folder.
CLR files must be in ACS format.
This function imports the cluster phenotype and abundance.
}
\details{
This function returns a 'Results' object including 'flowset', 'fcs.files' slots but not 'graph' and 'graph.layout' slots. 
The computation of marker range quantiles can be approximated using 'quantile.approximation' parameter which is more efficient in term of loading time and memory usage.
The name of each file will be used as the name of the sample. All FCS files corresponding to all the samples must be contained in a single folder.

For each cell, the selected cluster will be one having the greater probability higher than the 'prob.th' parameter.
}
