% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importResultsFromFCS}
\alias{importResultsFromFCS}
\title{Importation of clustering results from a folder containing FCS files}
\usage{
importResultsFromFCS(path, exclude.markers = c("cell_length", "FileNum",
  "density", "time"), clustering.markers = NULL, probs = c(0.05, 0.95),
  use.raw.medians = FALSE, quantile.approximation = FALSE,
  th.min_cells = 0, assignments = NULL)
}
\arguments{
\item{path}{a character specify the path of the FCS files}

\item{exclude.markers}{a character vector of markers to exclude (case insensitive)}

\item{clustering.markers}{a character vector specifying markers that have been used during the clustering procedure (if NULL, all markers will be considered as clustering markers)}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute marker range quantiles. First is the lower bound and second is the upper bound.}

\item{use.raw.medians}{a logical specifying if arcsinh transformed or raw medians will be used in the cluster expression matrix (FALSE by default)}

\item{quantile.approximation}{a logical specifying if marker range quantiles are computed using all cells (FALSE), or is the means of the quantile of each samples (TRUE)}

\item{th.min_cells}{a numeric specifying the minimum number of cell in a cluster of a sample to take its phenotype in account}

\item{assignments}{a data.frame containing all samples names (in rownames) and columns providing contextual associations like "bc" (biological conditions), "tp" (biological conditions) and "ind" (individuals)}
}
\value{
a S4 object of class 'Results'
}
\description{
The 'importResultsFromFCS()' function imports cell clustering results from FCS files contained in a specified folder.
This function imports the cluster phenotype matrix and count matrix.
This function apply an hyperbolic sine transformation to imported FCS data (unless 'use.raw.medians' = TRUE) and compute the marker range quantiles.
}
\details{
This function returns a 'Results' object including 'flowset', 'fcs.files' slots but not 'graph' and 'graph.layout' slots. 
The computation of marker range quantiles can be approximated using 'quantile.approximation' parameter which is more efficient in term of loading time and memory usage.
The name of each file will be used as the name of the sample. All FCS files corresponding to all the samples must be contained in a single folder. Importantly, FCS files must contain a channel called "cluster" allowing to identify which cluster is associated to each cell.
}
