% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importResultsFromSPADE}
\alias{importResultsFromSPADE}
\title{Importation of clustering results generated by SPADE}
\usage{
importResultsFromSPADE(path, exclude.markers = c("cell_length", "FileNum",
  "density", "time"), probs = c(0.05, 0.95), use.raw.medians = FALSE,
  quantile.approximation = FALSE, th.min_cells = 0, load.phenotype = TRUE,
  assignments = NULL)
}
\arguments{
\item{path}{a character specify the path of SPADE results folder}

\item{exclude.markers}{a character vector of markers to exclude (case insensitive)}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute marker range quantiles. First is the lower bound and second is the upper bound.}

\item{use.raw.medians}{a logical specifying if arcsinh transformed or raw medians will be used in the cluster expression matrix (FALSE by default)}

\item{quantile.approximation}{a logical specifying if marker range quantiles are computed using all cells (FALSE), or is the means of the quantile of each samples (TRUE)}

\item{th.min_cells}{a numeric specifying the minimum number of cell in a cluster of a sample to take its phenotype in account}

\item{load.phenotype}{a logical specifying if the phenotype matrix and fcs file will be loaded}

\item{assignments}{a data.frame containing all samples names (in rownames) and columns providing contextual associations like "bc" (biological conditions), "tp" (biological conditions) and "ind" (individuals)}
}
\value{
a S4 object of class 'Results'
}
\description{
The 'importResultsFromSPADE()' function imports SPADE cell clustering results from a specified path.
This function imports the cluster phenotype matrix and count matrix as well as the SPADE tree.
This function apply an hyperbolic sine transformation to imported FCS data (unless 'use.raw.medians' = TRUE) and compute the marker range quantiles.
}
\details{
This function returns a 'Results' object including 'flowset', 'fcs.files', 'graph' and 'graph.layout' slots. 
The computation of marker range quantiles can be approximated using 'quantile.approximation' parameter which is more efficient in term of loading time and memory usage.
}
