% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{importResultsFromTables}
\alias{importResultsFromTables}
\title{Importation of clustering results from text files}
\usage{
importResultsFromTables(cluster.abundances, cluster.phenotypes,
  clustering.markers = NULL, th.min_cells = 0, assignments = NULL)
}
\arguments{
\item{cluster.abundances}{a dataframe of cells abundances with clusters in row and samples in column}

\item{cluster.phenotypes}{a dataframe containing median marker expression values for each cluster of each sample. In additions of markers, the two first columns are are dedicated to "cluster" and "sample" identifiers.}

\item{clustering.markers}{a character vector specifying markers that have been used during the clustering procedure (if NULL, all markers will be considered as clustering markers)}

\item{th.min_cells}{a numeric specifying the minimum number of cell in a cluster of a sample to take its phenotype in account}

\item{assignments}{a data.frame containing all samples names (in rownames) and columns providing contextual associations like "bc" (biological conditions), "tp" (biological conditions) and "ind" (individuals)}
}
\value{
a S4 object of class 'Results'
}
\description{
The 'importResultsFromTables()' function imports cell clustering results from two dataframes ('cluster.abundances' and 'cluster.phenotypes').
This function returns a 'Results' object.
}
\details{
This function returns a 'Results' object without 'flowset', 'fcs.files', 'graph' and 'graph.layout' slots. 

The 'cluster.abundances' dataframe must be formatted with the cluster names in rownames as following:

    \tabular{ccccc}{
   x     \tab sample1 \tab sample2 \tab sample3 \tab ...\cr
   cluster1  \tab 749     \tab 5421    \tab   8424  \tab ...\cr
   cluster2  \tab 450     \tab 412     \tab   614   \tab ...\cr
   cluster3  \tab 288     \tab 782     \tab   478   \tab ...\cr
   ...       \tab ...     \tab ...     \tab   ...   \tab ...
}

The 'cluster.phenotypes' dataframe must be formatted as following:

   \tabular{cccccc}{
   sample     \tab cluster     \tab marker1     \tab marker2  \tab marker3 \tab ...\cr
   sample1  \tab cluster1 \tab 0.212    \tab 0.445   \tab 1.756   \tab ...\cr
   sample1  \tab cluster3 \tab 0.434    \tab 0.785   \tab 0.654   \tab ...\cr
   sample2  \tab cluster1 \tab 0.574    \tab 2.641   \tab 3.854   \tab ...\cr
   sample2  \tab cluster2 \tab 1.454    \tab 1.755   \tab -0.568  \tab ...\cr
   sample2  \tab cluster3 \tab 1.445    \tab 1.875   \tab -0.786  \tab ...\cr
   sample3  \tab cluster1 \tab 0.157    \tab 2.121   \tab 1.648   \tab ...\cr
   sample3  \tab cluster2 \tab 1.415    \tab 1.963   \tab 0.786   \tab ...\cr
   sample3  \tab cluster3 \tab 1.275    \tab 1.427   \tab 0.754   \tab ...\cr
   ...      \tab ...      \tab ...      \tab ...     \tab ...     \tab ...
}
}
