% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{importSPADEResults}
\alias{importSPADEResults}
\title{Import clustering results generated by SPADE}
\usage{
importSPADEResults(path, dictionary = data.frame(),
  exclude.markers = c("cell_length", "FileNum", "density", "time"),
  probs = c(0.05, 0.95), use.raw.medians = FALSE,
  quantile.approximation = FALSE, th.min_cells = 0)
}
\arguments{
\item{path}{a character specify the path of SPADE results folder}

\item{dictionary}{a two column dataframe providing the correspondence between the original marker names (first column) and the real marker names (second column)}

\item{exclude.markers}{a character vector of markers to exclude (case insensitive)}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute maker range quantiles. First is the lower bound and second is the upper bound.}

\item{use.raw.medians}{a logical specifying if arcsinh transformed or raw medians will be used in the cluster expression matrix (FALSE by default)}

\item{quantile.approximation}{a logical specifying if maker range quantiles are computed using all cells (FALSE), or is the means of the quantile of each samples (TRUE)}

\item{th.min_cells}{a numeric specifying the minimun number of cell in a cluster of a sample to take in account its phenotype}
}
\value{
a S4 object of class 'SPADEResults'
}
\description{
The 'importSPADEResults()' function imports SPADE cell clustering results from a specified path.
This function returns a 'SPADEResult' object.

This function import the expression matrix and count matrix as well as the SPADE tree.
This function apply an hyperbolic sine transformation to imported FCS data and compute the maker range quantiles.
}
\details{
The computation of maker range quantiles can be approximated using 'quantile.approximation' parameter which is more efficient in term of loading time and memory usage.
}

