% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mergeClusters}
\alias{mergeClusters}
\title{Merging of cell clusters from a Results object}
\usage{
mergeClusters(Results, clusters, name)
}
\arguments{
\item{Results}{a Results object}

\item{clusters}{a character vector containing the names of the clusters to remove}

\item{name}{a character specifiyng the name of the new cluster to create}
}
\value{
a Results object
}
\description{
This function is used to merge one or more clusters in a Results object.
}
\details{
The function merges the abundances and the phenotypes of clusters into a new one.
Clusters to merge are removed from the Results object.
}

