% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,AC-method}
\alias{names,CC-method}
\alias{names,DEC-method}
\alias{names,EnrichmentProfiles-method}
\alias{names,PhenoProfiles-method}
\alias{names,Results-method}
\alias{names,SPADEResults-method}
\title{Definition of class names}
\usage{
\S4method{names}{Results}(x)

\S4method{names}{SPADEResults}(x)

\S4method{names}{AC}(x)

\S4method{names}{DEC}(x)

\S4method{names}{CC}(x)

\S4method{names}{PhenoProfiles}(x)

\S4method{names}{EnrichmentProfiles}(x)
}
\arguments{
\item{x}{a SPADEVizR object}
}
\value{
a character providing the name of the object
}
\description{
Provides the name of each SPADEVizR object
}

