% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{PhenoProfiles-class}
\alias{PhenoProfiles}
\alias{PhenoProfiles-class}
\title{PhenoProfiles class definition}
\description{
PhenoProfiles is a S4 object containing the result of classifyPhenoProfiles() function.
It classify clusters by their markers properties provided by the fonction computePhenoTable()
}
\details{
PhenoProfiles is a printable and a plotable object calling the xxx() fonction
}
\section{Slots}{

\describe{
\item{\code{method}}{a character specifying the method used to classify cluster}

\item{\code{method.parameter}}{a list of parameters used by the selected method}

\item{\code{cluster.size}}{a numeric vector with the number of cell in each cluster (sum of all samples)}

\item{\code{cluster.number}}{a numeric providing the number of cluster}

\item{\code{class.number}}{a numeric providing the number of classes}

\item{\code{classes}}{a two column dataframe with the cluster in first colunm and corresponding classe in the second colunm}
}}

