% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,AC-method}
\alias{plot,CC-method}
\alias{plot,DEC-method}
\alias{plot,EnrichmentProfiles-method}
\alias{plot,PhenoProfiles-method}
\title{xxx}
\usage{
plot(x, ...)

\S4method{plot}{DEC}(x, ...)

\S4method{plot}{AC}(x, ...)

\S4method{plot}{CC}(x, ...)

\S4method{plot}{PhenoProfiles}(x, ...)

\S4method{plot}{EnrichmentProfiles}(x, ...)
}
\arguments{
\item{xxx}{}
}
\value{
xxx
}
\description{
xxx
}
\details{
xxx
}

