% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_prints.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,AC-method}
\alias{print,CC-method}
\alias{print,DEC-method}
\alias{print,EnrichmentProfiles-method}
\alias{print,PhenoProfiles-method}
\alias{print,Results-method}
\alias{print,SPADEResults-method}
\title{Textual preview for Results objects}
\usage{
\S4method{print}{Results}(x)

\S4method{print}{SPADEResults}(x)

\S4method{print}{AC}(x)

\S4method{print}{DEC}(x)

\S4method{print}{CC}(x)

\S4method{print}{PhenoProfiles}(x)

\S4method{print}{EnrichmentProfiles}(x)
}
\arguments{
\item{x}{a Results object}

\item{x}{a SPADEResults object}

\item{x}{a Abundant Clusters (AC) object}

\item{x}{a Differentially Enriched Clusters (DEC) object}

\item{x}{a Correlated Clusters (CC) object}

\item{x}{a PhenoProfiles object}

\item{x}{a EnrichmentProfiles object}
}
\value{
none

none

none

none

none

none

none
}
\description{
Prints a preview for a Results object.

Prints a preview for a SPADEResults object.

Prints a preview for a Abundant Clusters (AC) object.

Prints a preview for a Differentially Enriched Clusters (DEC) object.

Prints a preview for a Correlated Clusters (CC) object.

Prints a preview for a PhenoProfiles object.

Prints a preview for a EnrichmentProfiles object.
}

