% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{profilesViewer}
\alias{profilesViewer}
\title{profilesViewer}
\usage{
profilesViewer(profile.object, cluster.size = TRUE)
}
\arguments{
\item{profile.object}{a PhenoProfiles object or an EnrichmentProfiles object}

\item{cluster.size}{a logicial specifing if points size are related to cell count or not (TRUE by default)}

\item{show.unclassified}{a logical specifying if unclassified clusters must be shown or not (FALSE by default)}

\item{profile.object}{a PhenoProfiles object or an EnrichmentProfiles object}

\item{cluster.size}{a logicial specifing if points size are related to cell count or not (TRUE by default)}
}
\value{
a ggplot object

a ggplot object
}
\description{
Generate a graph representation of PhenoProfiles classes

Generate a graph representation of PhenoProfiles classes
}
\details{
xxx

xxx
}

