% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcSmallClusters}
\alias{qcSmallClusters}
\title{Computing of the fraction of clusters with low number of associated cells}
\usage{
qcSmallClusters(Results, clusters = NULL, th.size = 50,
  PDFfile = "qcSmallClusters.pdf", width = ncol(Results@cluster.abundances),
  height = nrow(Results@cluster.abundances)/4, tile.color = "black")
}
\arguments{
\item{Results}{a Results object}

\item{clusters}{a character vector containing the clusters names to be computed (by default all clusters will be computed)}

\item{th.size}{a numeric value specifying the minimum number of cells needed for a cluster to be considered a a small cluster}

\item{PDFfile}{a character specifying the location of the output PDF path}

\item{width}{a numeric specifying the plot width in the output PDF file}

\item{height}{a numeric specifying the plot height in the output PDF file}

\item{tile.color}{a character specifying the border color of the tiles (NA to remove tile borders)}
}
\value{
a list containing a numeric (perc numeric element) containing the fraction of clusters having a number of associated cells less than the threshold and a dataframe (small.clusters element) specifying the clusters having a number of associated cells less than the threshold
}
\description{
Computes the fraction of clusters having a number of associated cells less than a specific threshold (th.size parameter).
}
\details{
This function can also generate a PDF file summarizing the results.
}

