% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qcUniformClusters}
\alias{qcUniformClusters}
\title{Computing of the fraction of cluster with uniform phenotypes}
\usage{
qcUniformClusters(Results, clusters = NULL, only.clustering_markers = FALSE,
  uniform.test = "both", th.pvalue = 0.05, th.IQR = 2,
  density.PDFfile = "qcUniformClusters_density.pdf",
  density.PDFfile.dim = c(17, 10),
  heatmap.PDFfile = "qcUniformClusters_heatmap.pdf",
  heatmap.PDFfile.dim = c(length(Results@marker.names),
  length(Results@cluster.names)/4), tile.color = "black", verbose = TRUE,
  ...)
}
\arguments{
\item{Results}{a Results object (with 'flowset' slot not null)}

\item{clusters}{a character vector containing the clusters names to be computed (by default all clusters will be computed)}

\item{only.clustering_markers}{a logical specifying if only clustering marker densities must be displayed.}

\item{uniform.test}{a character specifying the qc assessment to perform}

\item{th.pvalue}{a numeric value specifying the pvalue threshold of the Hartigan's dip test (multimodal if p.value < th.pvalue)}

\item{th.IQR}{a numeric value specifying the IQR (interquartile range) threshold to assume a distribution as uniform}

\item{density.PDFfile}{a character specifying the output path of the marker expression density plots (or NULL to avoid this step)}

\item{density.PDFfile.dim}{a numeric vector specifying the width and the height of the density PDF file}

\item{heatmap.PDFfile}{a character specifying the output path of the marker expression accuracy heatmap (or NULL to avoid this step)}

\item{heatmap.PDFfile.dim}{a numeric vector specifying the width and the height of the heatmap PDF file}

\item{tile.color}{a character specifying the border color of the tiles (NA to remove tile borders)}

\item{verbose}{a boolean specifying if some messages must be displayed during the generation of the qc report}

\item{...}{supplemental parameters passed to the dip.test function}
}
\value{
a list containing a numeric (perc numeric element) specifying the percentage of clusters having only uniform clustering marker expressions and a dataframe (accuracy.matrix element) specifying if each marker of each cluster is uniform or not.
}
\description{
The `qcUniformClusters()` function allows to report as PDF files 2 kinds of information. 
Firstly, the identification of clusters having non-unimodal marker expression densities is performs using a Hartigan's dip test. 
Secondly, markers of cluster having a large spread of expressions (above a threshold) are identified.

If all the clustering marker expressions of a cluster checks this assessments (depending of the 'uniform.test' parameter), the cluster is considered as uniform.
}
\details{
The test of multimodal distribution is performs using a Hartigan's dip test. 
The null hypothesis of this test assume the distribution as unimodal 
Supplemental parameters passed to the dip.test function are ('simulate.p.value' and 'B').
'simulate.p.value' is a logical indicating whether to compute p-values by Monte Carlo simulation (FALSE by default).
'B' is an integer specifying the number of replicates used in the Monte Carlo test (2000 by default).

The spread of a marker expression is considered as high if the IQR (interquartile range) is higher than a provided threshold th.IQR.

The 'uniform.test' parameter can take 3 values: 
"unimodality": check if the distribution of the marker expression is unimodal or not using a Hartigans dip test
"spread": check if the spread of the marker expression is below an IQR threshold (interquantile range)
"both": check "unimodality" and "spread"
}
