% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{removeClusters}
\alias{removeClusters}
\title{Removing of cell clusters from a Results object}
\usage{
removeClusters(Results, clusters)
}
\arguments{
\item{Results}{a Results object}

\item{clusters}{a character vector containing the names of the clusters to remove}
}
\value{
a Results object
}
\description{
This function is used to remove one or more clusters from a Results object.
}

