% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{rename.markers}
\alias{rename.markers}
\title{Internal - Renaming cell markers}
\usage{
rename.markers(header, dictionary)
}
\arguments{
\item{header}{a character vector containing the original maker names}

\item{dictionary}{a character vector containing a correspondence between the original and the new marker names}
}
\value{
a character vector containing the renamed marker names
}
\description{
This function is used internally to rename the cell markers based on a dictionary.
}
\details{
dictionary is a data.frame used to rename the marker names. The first column must correspond to the original marker names, the second column must correspond to the new marker names.
}

