% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{Results-class}
\alias{Results}
\alias{Results-class}
\title{Results class definition}
\description{
Results is a S4 object containing cytometric results.
}
\details{
Results is the core (super classe) of SPADEResults object but allow importation of data comming from other software than SPADE. It aims to store the data comming from the importResults() function.
}
\section{Slots}{

\describe{
\item{\code{marker.expressions}}{a dataframe containing the number of cells for each cluster of each sample}

\item{\code{cells.count}}{a numerical dataframe containing marker median expressions for each cluster of each sample}

\item{\code{sample.names}}{a character vector containing the sample names}

\item{\code{marker.names}}{a character vector containing the markers names}

\item{\code{cluster.number}}{a numeric specyfing the number of clusters}
}}

