% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_prints.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,AC-method}
\alias{show,CC-method}
\alias{show,DEC-method}
\alias{show,EnrichmentProfiles-method}
\alias{show,PhenoProfiles-method}
\alias{show,Results-method}
\alias{show,SPADEResults-method}
\title{Textual previews for SPADEVizR objects}
\usage{
\S4method{show}{Results}(object)

\S4method{show}{SPADEResults}(object)

\S4method{show}{AC}(object)

\S4method{show}{DEC}(object)

\S4method{show}{CC}(object)

\S4method{show}{PhenoProfiles}(object)

\S4method{show}{EnrichmentProfiles}(object)
}
\arguments{
\item{object}{a SPADEVizR object}
}
\value{
none
}
\description{
Show a previews for a SPADEVizR objects.
}

