% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\docType{data}
\name{statSteamgraph}
\alias{statSteamgraph}
\title{title Internal - Transforms data for a steam graph (from by ggTimeSeries : https://github.com/Ather-Energy/ggTimeSeries)}
\format{An object of class \code{statSteamgraph} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
statSteamgraph
}
\arguments{
\item{xxx}{}
}
\value{
xxx
}
\description{
xxx
}
\details{
xxx
}
\keyword{datasets}

