% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{treeViewer}
\alias{treeViewer}
\title{Visualization of combined SPADE trees}
\usage{
treeViewer(SPADEResults, samples = NULL, stat.object = NULL,
  vertex_size = c(1, 15))
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{samples}{a named vector providing the correspondence between samples name (in rowname) and the logical value TRUE to use these samples for cell counting (all samples by default)}

\item{stat.object}{an AC, DEC or CC object to highligth significant clusters in the SPADE tree}

\item{vertex_size}{a numeric vector of two values indicating the range of}
}
\value{
a list of ggplot objects
}
\description{
Generates a tree representation showing combined SPADE trees.
}
\details{
xxx
}

