% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{unload.flowSet}
\alias{unload.flowSet}
\alias{unload.flowSet,Results-method}
\title{Unload 'flowSet' object from a 'Results' object}
\usage{
unload.flowSet(Results)

\S4method{unload.flowSet}{Results}(Results)
}
\arguments{
\item{Results}{a 'Results' object}
}
\value{
The new 'Results' object
}
\description{
This function unloads the 'flowSet' object in a 'Results' object.
}

