% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{unload.flowSet}
\alias{unload.flowSet}
\title{Unload 'flowSet' object from a 'SPADEResult' object}
\usage{
unload.flowSet(SPADEResult)
}
\arguments{
\item{SPADEResult}{a SPADEResult object (optional)}
}
\value{
The new 'SPADEResult' object
}
\description{
This function unloads the 'flowSet' object in a 'SPADEResult' object.
}

