% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcanoViewer}
\alias{volcanoViewer}
\title{Visualization of differentially abundant clusters}
\usage{
volcanoViewer(DAC = NULL, fc.log2 = TRUE, show.cluster.sizes = TRUE,
  show.all_labels = FALSE, show.on_device = TRUE)
}
\arguments{
\item{DAC}{an object of class 'DAC' (object returned by the 'computeDAC()' function)}

\item{fc.log2}{a logical specifying if fold-change or log2(fold-change) is use}

\item{show.cluster.sizes}{a logical specifying if dot sizes are proportional to cell counts}

\item{show.all_labels}{a logical specifying if all cluster labels must be show or just significant cluster}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generates a Volcano plot representation showing for each cluster
}
\details{
By default, only significant differentially abundant clusters are labeled. Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}

