% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{abundantClustersViewer}
\alias{abundantClustersViewer}
\title{Visualization of abundant clusters}
\usage{
abundantClustersViewer(AC, show.cluster_sizes = TRUE,
  show.all_labels = FALSE)
}
\arguments{
\item{AC}{an object of class AC (object returned by the 'computeAC()' function)}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to cell counts}

\item{show.all_labels}{a logical specifying if all cluster labels must be shown (or just significant clusters)}
}
\value{
a 'ggplot' object
}
\description{
Generates a scatter plot representation showing for each cluster its mean abundance and associated p-value.
}
\details{
By default, only significant abundant clusters are labeled. Labels for all clusters can be displayed by setting the 'show.all_labels' parameter to TRUE.
}

