% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{abundantClustersViewer}
\alias{abundantClustersViewer}
\title{abundantClustersViewer}
\usage{
abundantClustersViewer(AC, cluster.size = TRUE, all.label = FALSE)
}
\arguments{
\item{AC}{the AC object}

\item{cluster.size}{a logicial specifing if points size are related to cell count or not (TRUE by default)}

\item{all.label}{a logicial specifing if all cluster label must be show or just significant cluster}
}
\value{
a ggplot object
}
\description{
Generate a barplot representation to compare cluster abondances
}
\details{
By default, only abundant clusters are labeled, set parameter `all.label = TRUE` to visualize all labels.
}

