% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{abundantClustersViewer}
\alias{abundantClustersViewer}
\title{Visualization of abundant clusters}
\usage{
abundantClustersViewer(AC, show.cluster_sizes = TRUE,
<<<<<<< HEAD
  show.all_labels = FALSE, show.on_device = TRUE)
=======
  show.all_labels = FALSE)
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\arguments{
\item{AC}{an object of class AC (object returned by the 'computeAC()' function)}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to cell counts}

\item{show.all_labels}{a logical specifying if all cluster labels must be shown (or just significant clusters)}
<<<<<<< HEAD

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
=======
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\value{
a 'ggplot' object
}
\description{
Generates a scatter plot representation showing for each cluster its mean abundance and associated p-value.
}
\details{
By default, only significant abundant clusters are labeled. Labels for all clusters can be displayed by setting the 'show.all_labels' parameter to TRUE.
}

