% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{abundantClustersViewer}
\alias{abundantClustersViewer}
\title{Visualization of abundant clusters}
\usage{
abundantClustersViewer(AC, show.cluster_sizes = TRUE,
  show.all_labels = FALSE, show.on_device = TRUE,
  max.dots_size = max(AC@cluster.size))
}
\arguments{
\item{AC}{an object of class AC (object returned by the 'computeAC()' function)}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to number of associated cells}

\item{show.all_labels}{a logical specifying if all cluster labels must be shown. Only labels of significant clusters are displayed otherwise}

\item{show.on_device}{a logical specifying if the ggplot representation must be displayed on the device}

\item{max.dots_size}{a numeric specifying the number of associated cells in the largest dot}
}
\value{
a 'ggplot' object
}
\description{
Generates a scatter plot representation showing for each cluster: its mean abundance and associated p-value.

This representation displays the p-value (shown as -log10(p-value)) in the X-axis and the mean of cells abundance in the Y-axis, in a two-dimensional chart. 
Each dot in the representation corresponds to a cell cluster, and both p-value and mean thresholds are shown using red dashed lines. 
Abundant Clusters are highlighted in red and labeled. 
The size of the dots is proportional to the total number of associated cells in the considered samples.
}
\details{
By default, only significant abundant clusters are labeled. 
Labels for all clusters can be displayed by setting the 'show.all_labels' parameter to TRUE.
}

