% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{AC-class}
\alias{AC}
\alias{AC-class}
\title{Abundant Clusters (AC) class definition}
\description{
The AC object is a S4 object containing the identified of abundant cluster. 
It contains all informations about statistical parameters used to performs this test.
}
\details{
A cluster is considered as a significant abundant cluster if the 2 thresholds (mean and p-value) are reached.  
The AC object is returned by the computeAC() fonction.
The AC object owns methods to summuries main informations using print and show methods
Moreover this object could be exported as a tab separated file using the export method
}
\section{Slots}{

\describe{
\item{\code{sample.names}}{a character vector containing the samples used to compute the abundant clusters}

\item{\code{cluster.size}}{a numeric vector containing the number of cells ( -- sum of all samples -- ) for each cluster}

\item{\code{use.percentages}}{a logical specifying if computation was performed on percentage of cell abundance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the abundant clusters}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used (if any)}

\item{\code{th.mean}}{a numeric value specifying the mean threshold}

\item{\code{th.pvalue}}{a numeric value specifying the p-value threshold}

\item{\code{result}}{a data.frame containing for each cluster (first column): the mean (second column) and the standard deviation (third column) of the biological condition, the associated p-value (fourth column) and a logical (fifth column) specifying if the cluster is significantly abundant.}
}}

