% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{AC-class}
\alias{AC}
\alias{AC-class}
\title{Abundant Clusters (AC) class definition}
\description{
The 'AC' object is a S4 object containing the information related to the abundant clusters in a given biological condition. 
Moreover, this object contains parameters and results used in the statistical analysis.
}
\details{
A cluster is considered as a significant abundant cluster if its associated p-value and mean are below the specific thresholds 'th.pvalue'.  

The 'print()' and 'show()' can be used to display a summary of this object. 
Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'identifyAC()' function.
}
\section{Slots}{

\describe{
\item{\code{sample.names}}{a character vector containing the samples used to compute the abundant clusters}

\item{\code{cluster.size}}{a numeric vector containing the number of cells for each cluster}

\item{\code{use.percentages}}{a logical specifying if computation was performed on percentage of cell abundance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the abundant clusters}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used (if any)}

\item{\code{mu}}{a numeric specifying the theoretical value of the one sample statistical test}

\item{\code{th.pvalue}}{a numeric value specifying the p-value threshold}

\item{\code{results}}{a data.frame containing for each cluster (first column): the mean (second column) and the standard deviation (third column) of the biological condition, the associated p-value (fourth column) and a logical (fifth column) specifying if the cluster is significantly abundant.}
}}

