% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_viewers.R
\name{biplotViewer}
\alias{biplotViewer}
\title{biplotViewer}
\usage{
biplotViewer(SPADEResults, x.marker, y.marker, samples = NULL,
  clusters = NULL, sample.merge = FALSE, resample.ratio = NULL,
  show.on_device = TRUE)
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{x.marker}{a character indicating the marker name of the first dimension}

\item{y.marker}{a character indicating the marker name of the second dimension}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be used)}

\item{sample.merge}{a logical specifying if the selected samples must be merged in a single biplot}

\item{resample.ratio}{a numeric ratio (between 0 and 1) specifying the downsample ratio to show less dots (or NULL)}

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generates a biplot representation with two markers
}
\details{
In such representation, each dot corresponds to a cell profile and dots are ploted in a 2-dimentional space corresponding to the selected markers. 
When too cells dots are displayed, it can require some seconds. In order to seep up the computation, it is possible to reduce the number of cells displayed (downsampling) using the `resample.ratio` parameter.
}

