% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{biplotViewer}
\alias{biplotViewer}
\title{Visualization of marker co-expression}
\usage{
biplotViewer(Results, x.marker, y.marker, samples = NULL, clusters = NULL,
  resample.ratio = NULL, sample.merge = FALSE, show.on_device = TRUE,
  use.percentage = FALSE)
}
\arguments{
\item{Results}{a 'Results' object (with 'flowset' slot not null)}

\item{x.marker}{a character indicating the marker name of the first dimension}

\item{y.marker}{a character indicating the marker name of the second dimension}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be used)}

\item{resample.ratio}{a numeric ratio (between 0 and 1) specifying the down-sampling ratio to show less dots (or NULL)}

\item{sample.merge}{a logical specifying if the selected samples must be merged in a single biplot}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}

\item{use.percentage}{a logical specifying if the number of cells must be provided as percentages of parents}
}
\value{
a 'ggplot' object
}
\description{
The Biplot Viewer aims to visualize co-expressions between 2 markers using a biplot representation. 
In such representation, each cell is represented by a dot which is positioned in a two-dimensional space corresponding to the marker expressions.
}
\details{
Cells coming from specific clusters or samples can be selected using the `clusters` and `samples` parameters. 
Moreover, samples can be displayed independently (default behaviour) or merged.
In order to seed-up the computations, the number of cells to display can be down-sampled using the `resample.ratio` parameter.
}

