% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{boxplotViewer}
\alias{boxplotViewer}
\title{boxplotViewer}
\usage{
boxplotViewer(Results, conditions, clusters = NULL, use.percentages = TRUE,
  label = FALSE, violin = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{conditions}{conditions a named vector providing the correspondence between a sample name (in rownames) and the condition of this sample : NA to exclude a sample from tests}

\item{clusters}{a numerical vector containing the clusters to use in the representation}

\item{use.percentages}{a logical specifying if the visualisation should be performed on percentage}

\item{label}{a logical to show sample label or not (FALSE by default)}

\item{violin}{xxx}
}
\value{
a ggplot object
}
\description{
Generate a boxplot representation to compare the cell enrichment of biological conditions for each cluster
}
\details{
xxx
}

