% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Viewers.R
\name{boxplotViewer}
\alias{boxplotViewer}
\title{Visualization of cluster enrichment profiles conditions}
\usage{
boxplotViewer(Results, conditions, clusters = NULL, use.percentages = TRUE,
  show.legend = FALSE, show.violin = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{conditions}{conditions a named vector providing the correspondence between a sample name (in row names) and the condition of this sample : NA to exclude a sample from tests}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{use.percentages}{a logical specifying if the visualization must be performed on percentage}

\item{show.legend}{a logical specifying if the legend must be displayed}

\item{show.violin}{a logical specifying if the count distribution must be displayed}
}
\value{
a 'ggplot' object
}
\description{
Generate a boxplot representation displaying the cell abundance for each cluster. Clusters are gathered by given biological conditions.
}
\details{
Cells clusters are colored based on theirs associated biological samples.
}

