% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{boxplotViewer}
\alias{boxplotViewer}
\title{Visualization of cluster enrichment profiles conditions}
\usage{
boxplotViewer(Results, samples = NULL, clusters = NULL,
  use.percentages = TRUE, show.legend = TRUE, show.violin = TRUE,
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (required)}

\item{use.percentages}{a logical specifying if the visualization must be performed on percentage}

\item{show.legend}{a logical specifying if the legend must be displayed}

\item{show.violin}{a logical specifying if the count distribution must be displayed}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
The Boxplot Viewer aims to visualize and compare the abundances between several biological conditions for one single cluster or for a set of combined clusters.

This representation displays cell cluster abundances of each biological condition using boxplots.
Additionally, the abundance density of each biological condition can be visualize using violin representation.
Cluster abundance of each sample can also be visualized via colored dots.

Abundance from several clusters can be a by providing one or several cluster names to the `clusters` parameter.
Biological conditions information must be assigned to the samples in order to use this viewer (provided by the slot `assignments` in the `Results` object).
The cell cluster abundances could be displayed as percentages or absolute numbers using the `use.percentages` parameter (TRUE by default).
}
\details{
Biological conditions are specified in the Results object at the importation or using the 'assignContext()' function.
Cells clusters are colored based on theirs associated biological samples.
}
