% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{CC-class}
\alias{CC}
\alias{CC-class}
\title{Correlated Clusters (CC class) definition}
\description{
The 'CC' object is a S4 object containing coefficient of correlation associated between each cluster and a phenotypic variable.
Moreover this object contains all parameters used in the statistical analysis.
}
\details{
A cluster is considered as a significant correlated cluster if its associated p-value and correlation threshold are below the specific thresholds 'th.pvalue' and 'th.correlation'.  

The 'print()' and 'show()' can be used to display a summury of this object. Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'identifyCC()' function.
}
\section{Slots}{

\describe{
\item{\code{sample.names}}{a character vector containing the samples used to compute correlated clusters}

\item{\code{variable}}{a numeric vector containing the expression values of the associated variable}

\item{\code{cluster.size}}{a numeric vector containing number of cells ( -- sum of all samples -- ) for each cluster}

\item{\code{use.percentages}}{a logical specifying if computation was performed on percentage of cell abundance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the CC}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used (if any)}

\item{\code{th.correlation}}{a numeric value specifying the correlation threshold (R)}

\item{\code{th.pvalue}}{a numeric value specifying the p-value threshold}

\item{\code{result}}{a data.frame containing for each cluster (first column): the coefficiant of correlation R (second column) , the associated p-value (third column) and a logical (fourth column) specifying if the cluster is significantly correlated.}
}}

