% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{clusterViewer}
\alias{clusterViewer}
\title{ClusterViewer}
\usage{
clusterViewer(Results, samples = NULL, clusters = NULL, markers = NULL,
  show.mean = "both")
}
\arguments{
\item{Results}{a SPADEResuts or Result object (without quantiles bounds if a Results object is provided)}

\item{samples}{a named vector providing the correspondence between samples name (in rowname) and the logical value TRUE to use these samples (all samples by default)}

\item{clusters}{a character vector containing the clusters to use for the representation}

\item{markers}{a pattern describing markers to observe}

\item{show.mean}{a character : "none" "both" "only", "both" by default}
}
\value{
a list of ggplot objects
}
\description{
Cluster viewer
}
\details{
xxx
}

