% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeCC}
\alias{computeCC}
\title{Compute the correlation of SPADE cluster with a cynetics phenotype}
\usage{
computeCC(Results, variable, use.percentages = TRUE, method = "pearson",
  method.adjust = NULL, th.pvalue = 0.05, th.correlation = 0.75)
}
\arguments{
\item{Results}{a Results or SPADEResults object}

\item{variable}{a numerical named vector providing the correspondence between a sample name (in rowname) and the specific phenotype or NA to ignore a sample}

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character indicating the correlation method to use : "pearson", "spearman"}

\item{method.adjust}{a character specifying if the p-values should be corrected using multiple correction methods among : "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" and "fdr" (from 'stats::p.adjust' method)}

\item{th.pvalue}{a numeric specifying the p-value threshold (0.05 by default)}

\item{th.correlation}{a numeric specifying the absolute value of the correlation coefficient threshold (0.75 by default)}
}
\value{
a CC object
}
\description{
xxx
}

