% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeClique}
\alias{computeClique}
\title{Internal - Clique percolation classification}
\usage{
computeClique(data, clique.correlation.th = 0.7)
}
\arguments{
\item{data}{a numeric matrix with all clusters in rownames}

\item{clique.correlation.th}{a numeric value indicating the correlation coefficient threshold}
}
\value{
a dataframe containing for each cluster, its name and class
}
\description{
This function is used internally to classify clusters abundance profiles or phenotype profiles using a clique percolation algorithm.
}
\details{
This method first compute the Pearson correlation matrix and then use this matrix to generate an undirected graph.
In this graph, an edge is drawn between two nodes if the correlation coefficient in the adjacency matrix is above a specific threshold. 
This correlation threshold (a numeric double between 0 and 1 included, default is 0.7) provided using the 'clique.correlation.th' parameter.
After building the graph, the method looking for the largest cliques which are considered as classes of nodes. Cliques correspond to subgraph in which every two distinct vertices are adjacent.
}

