% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeClique}
\alias{computeClique}
\title{Internal - computeClique}
\usage{
computeClique(data, clique.correlation.th = 0.7)
}
\arguments{
\item{data}{a matrix with all clusters in rownames}

\item{eigencell.correlation.th}{}
}
\value{
a dataframe containing the cluster ID, the classe of this cluster and a clustering score.
}
\description{
xxx
}
\details{
xxx
}

