% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeClique}
\alias{computeClique}
\title{Internal - Clique percolation classification}
\usage{
computeClique(data, clique.correlation.th = 0.7)
}
\arguments{
\item{data}{a numeric matrix with all clusters in rownames}

\item{clique.correlation.th}{a numeric value indicating the correlation coefficient threshold}
}
\value{
a dataframe containing for each cluster, its name and class
}
\description{
This function is used internally to classify clusters enrichment profilies or phenotype profiles using a clique percolation algorithm.
}
\details{
xxx
}

