% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeEigenCellClusters}
\alias{computeEigenCellClusters}
\title{Internal - Eigen vector classification}
\usage{
computeEigenCellClusters(data, eigencell.correlation.th = 0.8)
}
\arguments{
\item{data}{a numeric matrix with all clusters in rownames}

\item{eigencell.correlation.th}{a numeric value indicating the correlation coefficient threshold}
}
\value{
a dataframe containing for each cluster, its name and class
}
\description{
This function is used internally to classify clusters enrichment profilies or phenotype profiles using eigen vector decomposition.
}
\details{
xxx
}

