% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{computeHierarchicalClustering}
\alias{computeHierarchicalClustering}
\title{Internal - Hierarchical classification}
\usage{
computeHierarchicalClustering(data, class.number = NULL,
  hierarchical.correlation.th = 0.8)
}
\arguments{
\item{data}{a matrix with all clusters in rownames}

\item{class.number}{a numeric specifying the number of classes}

\item{hierarchical.correlation.th}{a numeric value specifying the cut height}
}
\value{
a dataframe containing for each cluster, its name and class
}
\description{
This function is used internally to classify clusters abundance profiles or phenotype profiles using a hierarchical algorithm.
}
\details{
This function compute the Pearson correlation matrix associated to the provided matrix. 
It is to note that negative correlations are considered as uncorrelated.
This correlation matrix is used to performs a hierarchical classification.
If 'class.number' parameter is NULL, classification will be determined based on the cut height correlation threshold (i.e. 'hierarchical.correlation.th' parameter)
}

