% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeHierarchicalClustering}
\alias{computeHierarchicalClustering}
\title{Internal - computeHierarchicalClustering}
\usage{
computeHierarchicalClustering(data, class.number = NULL,
  hierarchical.correlation.th = 0.8)
}
\arguments{
\item{data}{a matrix with all clusters in rownames}

\item{class.number}{the number of classe to cluster, if class.number is NULL numer of classe will be determined base on hierarchical correlation threshold}

\item{hierarchical.correlation.th}{}
}
\value{
a dataframe containing the cluster ID, the classe of this cluster and a clustering score.
}
\description{
xxx
}
\details{
xxx
}

