% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_stats.R
\name{computeKMeans}
\alias{computeKMeans}
\title{Internal - computeKMeans}
\usage{
computeKMeans(data, k)
}
\arguments{
\item{data}{a matrix with all clusters in rownames}

\item{k}{number of classes}
}
\value{
a dataframe containing the cluster ID, the classe of this cluster and a clustering score.
}
\description{
xxx
}
\details{
xxx
}

