% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{correlatedClustersViewer}
\alias{correlatedClustersViewer}
\title{Visualization of correlated clusters}
\usage{
correlatedClustersViewer(CC, show.cluster.sizes = TRUE,
  show.all_labels = FALSE, show.on_device = TRUE)
}
\arguments{
\item{CC}{an object of class 'CC' (object returned by the 'computeCC()' function)}

\item{show.cluster.sizes}{a logical specifying if dot sizes are proportional to cell counts}

\item{show.all_labels}{a logical specifying if all cluster label must be show or just significant cluster}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generate a scatter plot representation showing for each cluster
}
\details{
By default, only significant correlated clusters are labeled. Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}

