% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{correlatedClustersViewer}
\alias{correlatedClustersViewer}
\title{Visualization of correlated clusters}
\usage{
correlatedClustersViewer(CC, show.cluster_sizes = TRUE,
  show.all_labels = FALSE, show.on_device = TRUE,
  max.dots_size = max(CC@cluster.size), varname = NULL)
}
\arguments{
\item{CC}{an object of class 'CC' (object returned by the 'computeCC()' function)}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to number of associated cells}

\item{show.all_labels}{a logical specifying if all cluster labels must be shown. Only labels of significant clusters are displayed otherwise}

\item{show.on_device}{a logical specifying if the ggplot representation must be displayed on the device}

\item{max.dots_size}{a numeric specifying the number of associated cells in the largest dot}

\item{varname}{a character indicating the name of the variable to display in the title}
}
\value{
a 'ggplot' object
}
\description{
Generates a volcano plot representation showing for each cluster: its correlation coefficient and associated p-value.

This representation displays the p-value (shown as -log10(p-value)) in the Y-axis and the correlation coefficient in the X-axis, in a two-dimensional chart.
Each dot in the representation corresponds to a cell cluster, and both correlation coefficient (positive and negative) and p-value thresholds are shown using red dashed lines. 
Correlated Clusters are highlighted in red and labeled. 
The size of dots is proportional to the total number of associated cells in the considered samples.
}
\details{
By default, only significant correlated clusters are labeled.
Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}
