% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_viewers.R
\name{countViewer}
\alias{countViewer}
\title{Visualization of cluster sizes}
\usage{
countViewer(Results, samples = NULL, clusters = NULL, min.cells = 0,
<<<<<<< HEAD
  sort = TRUE, show.samples = TRUE, show.on_device = TRUE)
=======
  sort = TRUE, show.samples = TRUE)
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{samples}{a named vector providing the correspondence between samples name (in row-names) and the logical value TRUE to use these samples (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{min.cells}{a numeric specifying the minimum number of cell (sum of all selected samples) to display a cluster}

\item{sort}{a logical specifying if clusters will be to be sorted (descending) based on the sum of all selected samples for each cluster}

\item{show.samples}{a logical specifying if the number of cells for all selected samples will be displayed}
<<<<<<< HEAD

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
=======
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\value{
a 'ggplot' object
}
\description{
Generate a two dimensional vizualisation showing the number of cells (sum of selected samples) of each cluster.
}

