% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{countViewer}
\alias{countViewer}
\title{Visualization of cluster sizes}
\usage{
countViewer(Results, samples = NULL, clusters = NULL, min.cells = 0,
  sort = TRUE, show.samples = TRUE, show.on_device = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{min.cells}{a numeric specifying the minimum number of cell (sum of all selected samples) to display a cluster}

\item{sort}{a logical specifying if clusters will be to be sorted (descending) based on the sum of all selected samples for each cluster}

\item{show.samples}{a logical specifying if the number of cells for all selected samples will be displayed}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Generate a two dimensional visualization showing the number of cells (sum of selected samples) of each cluster.
}

