% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{countViewer}
\alias{countViewer}
\title{Visualization of cluster sizes}
\usage{
countViewer(Results, samples = NULL, clusters = NULL, min.cells = 0,
  sort = TRUE, max.dots_size = NULL, show.samples = TRUE,
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{min.cells}{a numeric specifying the minimum number of cell (sum of all selected samples) to display a cluster}

\item{sort}{a logical specifying if the clusters will be to be sorted (descending) based on the sum of all selected samples for each cluster}

\item{max.dots_size}{a numeric specifying the number of cells in the largest dot}

\item{show.samples}{a logical specifying if the number of cells for all selected samples will be displayed}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
The Count Viewer aims to visualize the number of cells associated to each cluster. 
This representation displays the clusters in the X-axis and the total number of associated cells in the Y-axis.
Additionally, the numbers of cells associated to each cluster for each sample are also displayed using a jitter representation.
The size of the dots is proportional to the total number of associated cells.
}
\details{
By default, all clusters will be displayed but the representation can be restricted to a set of selected samples (using the `samples` parameter) or to a set of selected clusters (using the `clusters` parameter).
}
