% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{DEC-class}
\alias{DEC}
\alias{DEC-class}
\title{Differentially Enriched Clusters (DEC) class definition}
\description{
The DEC object is a S4 object containing the identified differentially enriched clusters beetween the selected conditions. 
It contains all informations about statistical parameters used to performs this test.
}
\details{
A cluster is considered as a significant differentially enriched cluster if the 2 thresholds (foldchange and p-value) are reached.  
The DEC object is returned by the computeDEC() fonction.
The AC object owns methods to summuries main informations using print and show methods
Moreover this object could be exported as a tab separated file using the export method
}
\section{Slots}{

\describe{
\item{\code{sample.cond1}}{a character specifying the names of the samples corresponding the first condition}

\item{\code{sample.cond2}}{a character specifying the names of the samples corresponding the second condition}

\item{\code{cluster.size}}{a numeric vector containing number of cells ( sum of all samples ) for each cluster}

\item{\code{use.percentages}}{a logical specifying if computations was performed on percentage of cell abondance}

\item{\code{method}}{a character containing the name of the statistical test used to identify the DEC}

\item{\code{method.adjust}}{a character containing the name of the multiple correction method used}

\item{\code{method.paired}}{a logical indicating if the test has beeen performed in a paired manner}

\item{\code{th.pvalue}}{a numeric vector with pvalue threshold}

\item{\code{th.fc}}{a numeric vector with foldchange threshold}

\item{\code{result}}{a dataframe with clusters in row. The first and second colunm contain the mean and the standard deviation for the first condition, the third and fourth the mean and standard deviation for the second condition, the sixth the fold-change and the seventh the pvalue and finnaly the significance according to pvalue threshold and fold change threshold}
}}

