% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{distogramViewer}
\alias{distogramViewer}
\title{Visualization of marker co-expressions}
\usage{
distogramViewer(Results, clusters = NULL, samples = NULL, markers = NULL)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{clusters}{a character vector containing the clusters names to be use (by default all clusters will be used)}

\item{samples}{a character vector specifying the samples to use}

\item{markers}{a character vector specifying the markers to be displayed}
}
\value{
a list of 'ggplot' objects
}
\description{
Generate a distogram representation showing the marker co-expressions.
}
\details{
xxx
}

