% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{distogramViewer}
\alias{distogramViewer}
\title{Distogram Viewer}
\usage{
distogramViewer(Results)
}
\arguments{
\item{Results}{a SPADEResults or Results object}
}
\value{
a list of ggplot objects
}
\description{
Generate a distogram representation showing the marker co-expression.
}
\details{
xxx
}

