% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{distogramViewer}
\alias{distogramViewer}
\title{biplotViewer}
\usage{
distogramViewer(Results)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{x.marker1}{a character indicating the marker name of the first dimension}

\item{y.marker2}{a character indicating the marker name of the second dimension}

\item{samples}{xxx}

\item{clusters}{xxx}

\item{default.min}{a numeric value indicating the lower bound of the biplot representation}

\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{x.marker1}{a character indicating the marker name of the first dimension}

\item{y.marker2}{a character indicating the marker name of the second dimension}

\item{samples}{xxx}

\item{clusters}{xxx}

\item{default.min}{a numeric value indicating the lower bound of the biplot representation}

\item{return.gg}{a logical indicating if the function should return a list of ggplot objects}

\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{x.marker1}{a character indicating the marker name of the first dimension}

\item{y.marker2}{a character indicating the marker name of the second dimension}

\item{samples}{xxx}

\item{clusters}{xxx}

\item{default.min}{a numeric value indicating the lower bound of the biplot representation}

\item{sample.merge}{xxx}

\item{resample.ratio}{xxx}
}
\value{
if return.gg is TRUE, the function returns a list of ggplot objects

if return.gg is TRUE, the function returns a list of ggplot objects

if return.gg is TRUE, the function returns a list of ggplot objects

a list of ggplot objects
}
\description{
Generates a biplot representation with two markers

Generates a biplot representation with two markers

Generates a biplot representation with two markers

Generate a distogram representation showing the marker co-expression.
}
\details{
In such representation, each dot corresponds to a cell profile and dot are ploted in a 2-dimentional space corresponding to the marker expressions.

In such representation, each dot corresponds to a cell profile and dot are ploted in a 2-dimentional space corresponding to the marker expressions.

In such representation, each dot corresponds to a cell profile and dot are ploted in a 2-dimentional space corresponding to the marker expressions.

xxx
}

