% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_viewers.R
\name{distogramViewer}
\alias{distogramViewer}
\title{Visualization of marker co-expressions}
\usage{
distogramViewer(Results, clusters = NULL, samples = NULL, markers = NULL)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{clusters}{a character vector containing the clusters names to be use (by default all clusters will be used)}

\item{samples}{a named vector providing the correspondence between samples name (in rowname) and the logical value TRUE to use these samples (all samples by default)}

\item{markers}{a character vector specifying the markers to be displayed}
}
\value{
a list of 'ggplot' objects
}
\description{
Generate a distogram representation showing the marker co-expressions.
}
\details{
A Pearson correlation matrix is calculated between selected markers using selected clusters and samples.
High positive correlated markers are shown by a green tile at their perpendicular intersection. 
In the same way, absence of correlation are shown by black tiles and negative correlation by red tiles.
}

