% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{distogramViewer}
\alias{distogramViewer}
\title{Visualization of marker co-expressions}
\usage{
distogramViewer(Results, samples = NULL, clusters = NULL, markers = NULL,
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be use (by default all clusters will be used)}

\item{markers}{a character vector specifying the markers to be displayed}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object.  This object also contains the correlations for each pair of marker (cor attribute).
}
\description{
The Distogram Viewer aims to visualize the pairwise co-expressions between all markers using a distogram representation. 
In such representation, each tile corresponds to the co-expression between 2 markers and is gradient-colored based on the Pearson correlation between the expressions of those 2 markers (as stored in the phenotype matrix).
Markers used as clustering markers are shown in blue.
}
\details{
The Pearson correlation is computed based on the marker expressions.
The visualization can be restricted to specific clusters, samples and markers by using respectively the `clusters`, `samples` and `markers` parameters.
}

