% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{EnrichmentProfiles-class}
\alias{EnrichmentProfiles}
\alias{EnrichmentProfiles-class}
\title{Classification of clusters based on their enrichment profiles (EnrichmentProfiles class) definition}
\description{
The 'EnrichmentProfiles' is a S4 object containing the information related to the cluster classification based on theirs enrichment profiles.


This object contains all information about the classification method and parameters used.
}
\details{
Five methods are available to classify cellular clusters: 'hierarchical_k', 'hierarchical_h', 'kmeans', 'eigencell' and 'clique'. Each method can paramaterized using the 'method.parameter' parameter.

The 'print()' and 'show()' can be used to display a summury of this object. Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'classifyEnrichmentProfiles()' function.
}
\section{Slots}{

\describe{
\item{\code{method}}{a character specifying the method used to classify cluster}

\item{\code{method.parameter}}{a numeric parameters associated with the chosen method}

\item{\code{cluster.size}}{a numeric vector with the number of cell in each cluster (sum of all samples)}

\item{\code{cluster.number}}{a numeric providing the number of cluster}

\item{\code{class.number}}{a numeric providing the number of classes}

\item{\code{classes}}{a two column dataframe with the cluster in first colunm and corresponding classe in the second colunm}
}}

