% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,AC-method}
\alias{export,AP-method}
\alias{export,CC-method}
\alias{export,DAC-method}
\alias{export,Results-method}
\title{Exportation of SPADEVizR objects}
\usage{
export(object, filename = "export.txt")

\S4method{export}{Results}(object, filename = "export.txt")

\S4method{export}{AC}(object, filename = "export.txt")

\S4method{export}{DAC}(object, filename = "export.txt")

\S4method{export}{CC}(object, filename = "export.txt")

\S4method{export}{AP}(object, filename = "export.txt")
}
\arguments{
\item{object}{a SPADEVizR object}

\item{filename}{a character indicating the location of the output file}
}
\value{
none
}
\description{
Exports a SPADEVizR object into a tab separated file.
}

