% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_axis}
\alias{g_axis}
\title{Internal - Extraction of ggplot axes}
\usage{
g_axis(gplot, x.axis = !y.axis, y.axis = !x.axis)
}
\arguments{
\item{gplot}{a 'ggplot' plot}

\item{x}{a logical value specifying if the x-axis must be extract}

\item{y}{a logical value specifying if the y-axis must be extract}
}
\value{
a 'ggplot' axis object
}
\description{
This function is used internally to extract axes from a 'ggplot' objet.
}
\details{
It is to note that 'x' and 'y' are mutuality excluded (both cannot be both TRUE) with priority to 'x'.
}

