% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_axis}
\alias{g_axis}
\title{title Internal - g_axis}
\usage{
g_axis(a.gplot, x = !y, y = !x)
}
\arguments{
\item{a.gplot}{a ggplot2 plot}

\item{x}{a logical, TRUE to extract x axis}

\item{y}{a logical, TRUE to extract y axis}
}
\value{
a ggplot2 axis
}
\description{
Extract axis from a plot
}
\details{
x and y are mutuality excluded (both can't be TRUE) with priority to x
}

