% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_legend}
\alias{g_legend}
\title{Internal - Extraction of ggplot legend}
\usage{
g_legend(gplot)
}
\arguments{
\item{a.gplot}{a ggplot2 plot}
}
\value{
a ggplot2 legend
}
\description{
This function is used internally to extract the legend from a 'ggplot' objet.
}

