% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{g_legend}
\alias{g_legend}
\title{Internal - Extraction of ggplot legend}
\usage{
g_legend(gplot)
}
\arguments{
\item{gplot}{a 'ggplot' plot}
}
\value{
a 'ggplot' legend object
}
\description{
This function is used internally to extract the legend from a 'ggplot' object.
}

