% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_reports.R
\name{generateReport}
\alias{generateReport}
\title{Generate report and plot.}
\usage{
generateReport(Results, PDFfile, reports = c("pheno", "kinetic", "cluster",
  "kinetic_cluster", "tree", "disto", "stream", "MDS"), clusters = NULL,
  markers = NULL, assignments = NULL, stat.objects = list(),
  profile.objects = list(), width = 29.7, height = 21)
}
\arguments{
\item{Results}{a SPADEResuts or Result object}

\item{PDFfile}{a character specifying the output path}

\item{reports}{a vector of the plot names to add in the report (following the vector order). Plot names are : "pheno", "MDS", "cluster", "tree", "disto","stream". By default all plots will be add}

\item{clusters}{a character vector of ID cluster to be reported}

\item{markers}{a character vector of markers to be reported}

\item{stat.objects}{a list of stat.object (object of type DEC, AC or CC)}

\item{profile.objects}{a list of profile.objects (object of type PhenoProfiles or EnrichmentProfiles)}

\item{width}{a numeric specifying the plot width}

\item{height}{a numeric specifying the plot height}
}
\description{
Generate a report based on the SPADE result object.
}

