% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_reports.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report including SPADEVizR plots.}
\usage{
generateReport(Results, PDFfile = "report.pdf", select.plots = c("count",
  "heatmap", "tree", "disto", "MDSclusters", "pheno"), clusters = NULL,
  markers = NULL, samples = NULL, assignments = NULL, conditions = NULL,
  stat.objects = list(), width = 50, height = 30, verbose = TRUE)
}
\arguments{
\item{Results}{a 'SPADEResults' or 'Result' object}

\item{PDFfile}{a character specifying the output path}

\item{select.plots}{a vector combining character and stat objects ('AC', 'DAC', 'CC' and 'CCR) specifying the order of the desired plots (see details)}

\item{clusters}{a character vector of clusters to include in the report (all will be included by default)}

\item{markers}{a character vector of markers to include in the report (all will be included by default)}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{assignments}{a 2 column data.frame with the samples names in row names providing firstly the time-points (numeric) and secondly the individuals (character) of the experiment}

\item{conditions}{conditions a named vector providing the correspondence between a sample name (in row names) and the condition of this sample or NA to exclude}

\item{width}{a numeric specifying the plot width in centimeter}

\item{height}{a numeric specifying the plot height in centimeter}
}
\description{
Generate a customizable PDF report based on SPADEVizR vizualisation features.
Available plots are :
\itemize{
\item {"count" (included by default):}{Display an representation showing the number of cells for each cluster}
\item {"tree" (included by default):}{Display a tree representation showing combined SPADE trees}
\item {"heatmap" (included by default):}{Display an heatmap representation}
\item {"boxplot":}{Display a boxplot representation. This plot required to provide the 'conditions' parameter}
\item {"kinetics":}{Display a kinetic representation for each cluster. This plot required to provide the 'assignments' parameter}
\item {"stream":}{Display a streamgraphViewer representation showing the evolution of cells abundance. The 'clusters' parameter is required}
\item {"pheno" (included by default):}{Display a parallel coordinate representation showing for each cluster the marker median expression}
\item {"MDSclusters" (included by default):}{Display the cluster similarities using MDS}
\item {"MDSsamples":}{Display the samples similarities using MDS}
\item {"disto" (included by default):}{Display a distogram representation showing the marker co-expressions}
\item {"kinetics_pheno":}{Display a kinetic representation and a parallel coordinate juxtaposed (are arranged one on the side of the other) for each cluster}
\item {"boxplot_pheno":}{Display a boxplot representation and a parallel coordinate juxtaposed (are arranged one on the side of the other) for each cluster}
\item {AC, DAC, CC and CCR objects}
}
}

