% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_reports.R
\name{generateReport}
\alias{generateReport}
\title{Generate a report including SPADEVizR plots.}
\usage{
generateReport(results, PDFfile, plots.names = c("pheno", "cluster", "tree",
  "disto", "stream", "MDSclusters"), clusters = NULL, markers = NULL,
  assignments = NULL, conditions = NULL, stat.objects = list(),
  profile.objects = list(), width = 29.7, height = 21)
}
\arguments{
\item{PDFfile}{a character specifying the output path}

\item{plots.names}{a character vector specifying the names (see details) and the order of the desired plots}

\item{clusters}{a character vector of clusters to include in the report (all will be included by default)}

\item{markers}{a character vector of markers to include in the report (all will be included by default)}

\item{assignments}{a 2 column data.frame with the samples names in row names providing firstly the time-points (numeric) and secondly the individuals (character) of the experiment}

\item{stat.objects}{a vector of stat.object to be displayed in the report (object of class 'DEC', 'AC' or 'CC')}

\item{profile.objects}{a vector of profile.objects to be displayed in the report (object of class 'PhenoProfiles' or 'EnrichmentProfiles')}

\item{width}{a numeric specifying the plot width}

\item{height}{a numeric specifying the plot height}

\item{Results}{a 'SPADEResults' or 'Result' object}
}
\description{
Generate a customizable PDF report based on SPADEVizR vizualisation features.
Available plots are :
\itemize {
\item "pheno" (included by default): Display an heatmap representation
\item "kinetic": Display a kinetic representation for each cluster. This plot required to provide the 'assignment' parameter.
\item "cluster" (included by default): Display a parallel coordinate representation showing for each cluster the marker median expression.
\item "boxplot": Display a boxplot representation. This plot required to provide the 'conditions' parameter.
\item "kinetic_cluster": Display a kinetic representation and a parallel coordinate juxtaposed (are arranged one on the side of the other) for each cluster.
\item "boxplot_cluster": Display a boxplot representation and a parallel coordinate juxtaposed (are arranged one on the side of the other) for each cluster.
\item "tree" (included by default): Display a tree representation showing combined SPADE trees. 
\item "disto" (included by default): Display a distogram representation showing the marker co-expressions.
\item "stream" (included by default): Display a 
\item "MDSclusters" (included by default): Display a Multidimensional Scaling (MDS) representation showing the
\item "MDSsamples": Display a Multidimensional Scaling (MDS) representation showing the. This plot required to provide the 'assignment' parameter.
}
}

