% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{generateReport}
\alias{generateReport}
\title{Generating SPADEVizR reports}
\usage{
generateReport(Results, PDFfile = "report.pdf", select.plots = c("count",
  "heatmap", "MDSclusters", "pheno"), clusters = NULL, markers = NULL,
  samples = NULL, stat.objects = list(), width = 30, height = 10,
  verbose = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{PDFfile}{a character specifying the output path}

\item{select.plots}{a vector combining character and stat objects ('AC', 'DAC', 'CC' and 'AP') specifying the order of the desired plots (see details)}

\item{clusters}{a character vector of clusters to include in the report (all will be included by default)}

\item{markers}{a character vector of markers to include in the report (all will be included by default)}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{stat.objects}{a list containing one or several AC, DEC, CC or AP objects to plot in the report}

\item{width}{a numeric specifying the width of the PDF file}

\item{height}{a numeric specifying the height of the PDF file}

\item{verbose}{a boolean specifying if some messages must be displayed during the generation of the report}
}
\value{
none
}
\description{
Generates a customizable PDF report based on SPADEVizR visualization features.
}
\details{
Available plots are :
\itemize{
\item {"count" (included by default):}{Displays a Count Viewer representation;}
\item {"tree":}{Displays a SPADE Tree Viewer representation;}
\item {"heatmap" (included by default):}{Displays an Heatmap Viewer representation;}
\item {"boxplot":}{Displays a Boxplot Viewer representation;}
\item {"kinetics":}{Displays a Kinetic Viewer representation;}
\item {"streamgraph":}{Display a Streamgraph Viewer representation;}
\item {"pheno" (included by default):}{Displays a Pheno Viewer representation;}
\item {"MDSclusters" (included by default):}{Displays a MDS Viewer representation at the cluster level}
\item {"MDSsamples":}{Displays a MDS Viewer representation at the sample level}
\item {"distogram":}{Displays a Distogram Viewer representation;}
\item {"kinetics_pheno":}{Displays two "kinetics" and "cluster" representations juxtaposed (one on the side of the other) for each cluster;}
\item {"boxplot_pheno":}{Displays two "boxplot" and "cluster" representations juxtaposed (one on the side of the other) for each cluster;}
\item {AC, DAC, CC and AP objects}
}
}

