% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggheatmap.plot}
\alias{ggheatmap.plot}
\title{Internal - Generate an heatmap by assembling elements}
\usage{
ggheatmap.plot(list, col.width = 0.15, row.width = 0.15)
}
\arguments{
\item{list}{the list of ggplot object provided by ggheatmap}

\item{col.width}{size of horizontal dendrogram}

\item{row.width}{size of vertical dendrogram}
}
\value{
a ggplot2 axis
}
\description{
This function is used internally to displays the heatmap elements build by 'ggheatmap()'
}

