% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_utils.R
\name{ggheatmap}
\alias{ggheatmap}
\title{title Internal - Create a list of elements allowing to build a heatmap}
\usage{
ggheatmap(mat, dendrogram.type = "rectangle", num = 5,
  dists = c("eucledian", "eucledian"))
}
\arguments{
\item{mat}{a matrix}

\item{dendrogram.type}{a caracter spycifing the look of dendrograms ("rectangle" or "triangle", "rectangle" by default)}

\item{num}{xxx}
}
\value{
a list of 3 plots (top dendrogram, right dendrogram, heatmap)
}
\description{
Use the function ggheatmap.plot to display this heatmap
}
\details{

}

