% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{heatmapViewer}
\alias{heatmapViewer}
\title{Visualization of all clusters phenotypes using categorical heatmap}
\usage{
heatmapViewer(Results, clusters = NULL, markers = NULL, num = 5,
  clustering = "both", tile.color = "black", show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{clusters}{a character vector providing the clusters to be used (all clusters by default)}

\item{markers}{a character vector providing the markers to be used (all markers by default)}

\item{num}{a numeric value specifying the number of markers expression categories to be used}

\item{clustering}{a character specifying which clustering must be build ("markers", "clusters", "both" or "none")}

\item{tile.color}{a character specifying the border color of the tiles (NA to remove tile borders)}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object. This object also contains the categorical phenotipical heatmap (pheno.table attribute) and the marker and cluster hierarchical clustering dendrograms (markers.hc and clusters.hc attributes)
}
\description{
The Heatmap Viewer aims to visualize the phenotypes of all cell clusters.
This representation displays marker expressions of all clusters using a categorical heatmap (5 categories are computed by default). 
The range expression of each cell marker is discretized in several categories between bounds of marker expressions.
Each marker of each cluster is then categorized into one category based on the mean of median marker expressions.
Markers used as clustering markers are shown in blue.
}
\details{
Both cell clusters and cell markers can be clustered using a hierarchical clustering (using the `dendrograms`, generating marker and cluster dendrograms by default).

It is to note that the Heatmap Viewer is the default viewer for `Results` objects.

The structures of the marker and cluster hierarchical clustering dendrograms are also returned by this fucntion.
}

