% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{identifyAC}
\alias{identifyAC}
\title{Identification of the Abundant Clusters}
\usage{
identifyAC(Results, samples, use.percentages = TRUE, method = "t.test",
  method.adjust = NULL, mu = 0, th.pvalue = 0.05)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used}

\item{use.percentages}{a logical specifying if the computations should be performed on percentage}

\item{method}{a character specifying the statistical method used to identify the Abundant Clusters. The parameter can take the values "t.test" or "wilcox.test"}

\item{method.adjust}{a character specifying if the p-values should be corrected using multiple correction methods among: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" and "fdr" (from 'stats::p.adjust' method)}

\item{mu}{a numeric specifying the theoretical value (mu) of the one sample statistical test}

\item{th.pvalue}{a numeric specifying the p-value threshold}
}
\value{
a S4 object of class 'AC'
}
\description{
This function is used to identify the Abundant Clusters (AC), that is to say clusters having cell abundance (absolute or relative) statistically greater than a specific threshold.
Abundant Clusters are identified using a one-sample test (parametrized or non-parametrized).
P-values can be corrected for multiple comparisons.
}
