% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{importResults}
\alias{importResults}
\title{Import clustering results generated by other algorithms}
\usage{
importResults(cells.count, marker.expressions)
}
\arguments{
\item{cells.count}{a dataframe of cells abondances with clusters in row and samples in column}

\item{marker.expressions}{a dataframe containing median marker expression values for each cluster of each sample. In additions of markers, the 2 two first columns are are dedicated to "cluster" and "sample"}
}
\value{
Results a result object
}
\description{
The 'importX()' function imports cell clustering results from two dataframes ('cells.count' and 'marker.expressions').
This function returns a 'Result' object.
}
\details{
The detailled format of 'cells.count' dataframe must be formated as following:

\tabular{ccc}{
   cluster \tab sample1 \tab sample1\cr
   cluster1 \tab 749 \tab 5421\cr
   cluster2 \tab 450 \tab 412\cr
}

The detailled format of 'marker.expressions' dataframe must be formated as following:

\tabular{cccc}{
   sample \tab cluster \tab marker1 \tab marker2\cr
   sample1 \tab cluster1 \tab 0.2 \tab 0.3\cr
   sample1 \tab cluster2 \tab 0.1 \tab 0.3\cr
   sample2 \tab cluster1 \tab 0.5 \tab 2.3\cr
   sample2 \tab cluster2 \tab 1   \tab 1.3\cr
}
}

