% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{importResults}
\alias{importResults}
\title{Import clustering results generated by other algorithms}
\usage{
importResults(cluster.abundances, cluster.phenotypes, th.min_cells = 50)
}
\arguments{
\item{cluster.abundances}{a dataframe of cells abundances with clusters in row and samples in column}

\item{cluster.phenotypes}{a dataframe containing median marker expression values for each cluster of each sample. In additions of markers, the 2 two first columns are are dedicated to "cluster" and "sample"}

\item{th.min_cells}{a numeric specifying the minimun number of cell in a cluster of a sample to take in account its phenotype}
}
\value{
a S4 object of class 'Results'
}
\description{
The 'importResult()' function imports cell clustering results from two dataframes ('cells.count' and 'marker.expressions').
This function returns a 'Result' object.
}
\details{
The 'cluster.abundances' dataframe must be formated with the cluster names in rownames as following:

\tabular{ccc}{
   X \tab sample1 \tab sample1\cr
   cluster1 \tab 749 \tab 5421\cr
   cluster2 \tab 450 \tab 412\cr
}

The 'cluster.phenotypes' dataframe must be formated as following:

\tabular{cccc}{
   sample \tab cluster \tab marker1 \tab marker2\cr
   sample1 \tab cluster1 \tab 0.2 \tab 0.3\cr
   sample1 \tab cluster2 \tab 0.1 \tab 0.3\cr
   sample2 \tab cluster1 \tab 0.5 \tab 2.3\cr
   sample2 \tab cluster2 \tab 1   \tab 1.3\cr
}
}

