% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_imports.R
\name{importSPADEResults}
\alias{importSPADEResults}
\title{importSPADEResults}
\usage{
importSPADEResults(path, dictionary = data.frame(), exclude.markers = NULL,
  probs = c(0.05, 0.95), use.raw.medians = FALSE,
  quantile.heuristic = FALSE)
}
\arguments{
\item{path}{the a character specify the path of SPADE results folder}

\item{dictionary}{a two column data.frame providing the correspondance between the original marker names (first column) and the new marker names (second column)}

\item{exclude.markers}{a list of markers to exclude}

\item{probs}{a vector of probabilities with 2 values in [0,1] to compute quantiles. First is the lower bound and second is the lower bound.}

\item{use.raw.medians}{a logicial specifying if "transformed" or "raw" medians will be use (FALSE by default)}

\item{quantile.heuristic}{a logicial specifying if quantile are compute for with all cells (FALSE), or is the means of the quantile of each samples (TRUE)}
}
\value{
SPADEResults a SPADE result object
}
\description{
Import the SPADE results from a specified path to a SPADEResult object.

Generate a SPADEresults object based on SPADE results.

The tables are loaded, the SPADE tree is loaded as well as the layout
}
\details{
We advice to use quantile.heuristic = TRUE because it is faster and much more efficient in term of memory with a low impact on precision.
}

