% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{kineticsViewer}
\alias{kineticsViewer}
\title{Visualization of cluster enrichment profiles kinetics}
\usage{
kineticsViewer(Results, assignments, clusters = NULL,
  use.percentages = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{assignments}{a 2 column data.frame with the samples names in rownames providing firstly the timepoints (numeric) and secondly the individuals (caracter) of the experiment}

\item{clusters}{a character vector containing the clusters names to be vizualised (by default all clusters will be displayed)}

\item{use.percentages}{a logical specifying if the visualisation should be performed on percentage}
}
\value{
a 'ggplot' object
}
\description{
Generates a kinetics plot representation showing for each cluster its enrichment profiles at each timepoint of each individual.
}
\details{
xxx
}

