% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{kineticsViewer}
\alias{kineticsViewer}
\title{Visualization of cluster enrichment profiles kinetics}
\usage{
kineticsViewer(Results, assignments, clusters = NULL,
  use.percentages = TRUE, show.on_device = TRUE, verbose = FALSE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{assignments}{a 2 column data.frame with the sample names in row names providing firstly the time-points (numeric) and secondly the individuals (character) of the experiment}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{use.percentages}{a logical specifying if the visualization should be performed on percentage}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}

\item{verbose}{a logical specifying if the details of computation must be printed}
}
\value{
a 'ggplot' object
}
\description{
Generates a kinetics plot representation showing for each cluster its enrichment profiles at each time-point of each individual.
}
\details{
Time-points are sorted in a way that strings with embedded numbers are in the correct order
}

