% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{kineticsViewer}
\alias{kineticsViewer}
\title{Visualization of cluster enrichment profiles kinetics}
\usage{
kineticsViewer(Results, samples = NULL, clusters = NULL,
  use.percentages = TRUE, show.on_device = TRUE, scale_y = NULL)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character or a character vector containing the cluster to be shown or if multiple clusters which will be merged}

\item{use.percentages}{a logical specifying if the visualization should be performed on percentage}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}

\item{scale_y}{a numeric value specifying the maximal value in y axsis}
}
\value{
a 'ggplot' object
}
\description{
The Kinetics Viewer aims to visualize the cell cluster abundances in a kinetics manner. 
This representation displays the cell abundances over the time for each individual. 
The cell cluster abundances could be displayed as percentages or absolute numbers using the `use.percentages` parameter (TRUE by default).
}
\details{
Timepoints and indivuals information must be assigned to the samples in order to use this viewer (provided by the slot `assignments` in the `Results` object).
}
