% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{load.flowSet}
\alias{load.flowSet}
\title{Loading of FCS files object into a 'Results' object}
\usage{
load.flowSet(Results = NULL, fcs.files, exclude.markers, use.raw.medians,
  pattern = ".fcs.density.fcs.cluster.fcs")
}
\arguments{
\item{Results}{a Results object (with 'fcs.files' slot not null) (optional)}

\item{fcs.files}{a character vector containing the absolute path of the original FCS files}

\item{exclude.markers}{a character vector of markers to exclude (case insensitive)}

\item{use.raw.medians}{a logical specifying if the arcsinh transformation must be performed or not}

\item{pattern}{a character specifying the pattern of the FCS file}
}
\value{
a S4 'flowSet' object
}
\description{
This function loads the FCS files to the 'flowset' slot of the 'Results' object.
}
\details{
If a 'Results' object is provided, others parameters ('fcs.files', 'exclude.markers' and 'use.raw.medians') will be ignored (they will be retrieved from the 'Results' object).
}

