% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{MDSViewer}
\alias{MDSViewer}
\title{MDS viewer.}
\usage{
MDSViewer(Results, use.percentages = TRUE, assignments, clusters = NULL,
  space = "clusters", dist.method = "euclidean")
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{use.percentages}{a logical specifying if the visualisation should be performed on percentage}

\item{assignments}{a 2 column data.frame with the samples names in rownames providing firstly the timepoints (numeric) and secondly the individuals (caracter) of the experiment}

\item{clusters}{specify the set of clusters to use}

\item{space}{a caracter specifying the space ("clusters" or "samples", cluster by default)}

\item{dist.method}{a character string containing the name of the distance measure to use}
}
\value{
a list of ggplot objects
}
\description{
Generate a MDS representation based on the SPADE result object.
}
\details{
xxx
}

