% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{MDSViewer}
\alias{MDSViewer}
\title{Visualization of SPADE cluster or sample similarities using MDS}
\usage{
MDSViewer(Results, use.percentages = TRUE, assignments = NULL,
  clusters = NULL, space = "clusters", dist.method = "euclidean",
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{use.percentages}{a logical specifying if the visualization should be performed on percentage}

\item{assignments}{a 2 column data.frame with the samples names in row names providing firstly the biological condition and secondly the individuals of the experiment}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{space}{a character specifying the space ("clusters" or "samples", "cluster" by default)}

\item{dist.method}{a character string containing the name of the distance measure to use}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a list of 'ggplot' objects
}
\description{
Generate a Multidimensional Scaling (MDS) representation showing the similarities between SPADE results based on theirs abundances.
}
\details{
The 'space' parameter specifying if the cluster or sample similarities will be determined using MDS.
Available method for the 'dist.method' parameter are : "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski"
}

