% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{MDSViewer}
\alias{MDSViewer}
\title{Visualization of SPADE cluster or sample similarities using MDS}
\usage{
MDSViewer(Results, space = "clusters", samples = NULL, clusters = NULL,
  use.percentages = TRUE, dist.method = "euclidean",
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{space}{a character specifying the space ("clusters" or "samples", "cluster" by default)}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{use.percentages}{a logical specifying if the visualization should be performed on percentage}

\item{dist.method}{a character string containing the name of the distance measure to use}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
Multidimensional Scaling (MDS) methods aim to represent the similarities and differences among high-dimensional objects into a space of a lower number of dimensions, generally in two or three dimensions for visualization purposes.
In MDS representations, the Kruskal Stress (KS) indicates the percentage of information lost during the dimensionality reduction process.

The MDS Viewer aims to visualize the similarities between samples or clusters based on their abundances. 
In such representation, each dot represents a sample or a cluster and the distances between the dots are proportional to the Euclidean distance between these objects.

The representation space can be specified using the `space` parameter  ("samples" or "clusters").
}
\details{
In the case of "samples" space, biological conditions information can be assigned to the samples (provided by the slot `assignments` in the `Results` object).
This parameter must be a dataframe with sample names in row names and 2 other columns specifying the biological conditions and individuals.
}
