% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,AC-method}
\alias{names,CC-method}
\alias{names,DEC-method}
\alias{names,EnrichmentProfiles-method}
\alias{names,PhenoProfiles-method}
\alias{names,Results-method}
\alias{names,SPADEResults-method}
\title{Class name}
\usage{
\S4method{names}{Results}(x)

\S4method{names}{SPADEResults}(x)

\S4method{names}{AC}(x)

\S4method{names}{DEC}(x)

\S4method{names}{CC}(x)

\S4method{names}{PhenoProfiles}(x)

\S4method{names}{EnrichmentProfiles}(x)
}
\arguments{
\item{x}{a SPADEVizR object}

\item{x}{a SPADEResults object}

\item{x}{a AC object}

\item{x}{a DEC object}

\item{x}{a CC object}

\item{x}{a PhenoProfiles object}

\item{x}{a EnrichmentProfiles object}
}
\value{
a character specifying the name of the object

none

none

none

none

none

none
}
\description{
Provides the name of each SPADEVizR object

Named a SPADEResults object.

Named a AC object.

Named a DEC object.

Named a CC object.

Named a PhenoProfiles object.

Named a EnrichmentProfiles object.
}

