% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{PhenoProfiles-class}
\alias{CCR}
\alias{PhenoProfiles-class}
\title{Classification of clustering Results (CCR class) definition}
\description{
The 'CR' is a S4 object containing the information related to the cluster classification based on theirs marker expressions.


This object contains all information about the classification method and parameters used.
}
\details{
Five methods are available to classify cellular clusters: 'hierarchical_k', 'hierarchical_h', 'kmeans', 'eigencell' and 'clique'. Each method can parameterized using the 'method.parameter' parameter.
 
The 'print()' and 'show()' can be used to display a summary of this object. Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'classifyPhenoProfiles()' function.
}
\section{Slots}{

\describe{
\item{\code{type}}{a character specifying if the classification is based on the phenotype profiles or on the enrichment profiles}

\item{\code{class.number}}{a numeric value specifying the number of clusters}

\item{\code{method}}{a character specifying the method used to classify cluster}

\item{\code{method.parameter}}{a named list of parameters used by the classification method}

\item{\code{classes}}{a two column dataframe with the cluster in first colunm and corresponding classe in the second colunm}
}}

