% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{phenoViewer}
\alias{phenoViewer}
\title{Pheno Viewer}
\usage{
phenoViewer(SPADEResults, pheno.table = NULL, num = 5)
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{pheno.table}{a result of the function computePhenoTable}

\item{num}{a numeric indicating the precision of computed expression scores}
}
\value{
a list of ggplot objects
}
\description{
Heatmap Viewer aims to xxx
}
\details{
xxx
}

