% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Viewers.R
\name{phenoViewer}
\alias{phenoViewer}
\title{Visualization of all clusters phenotypes as an heatmap}
\usage{
phenoViewer(SPADEResults, num = 5)
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{num}{a numeric value specifying the number of markers expression categories to use}
}
\value{
a list of ggplot objects
}
\description{
Generates an heatmap representation showing for all clusters the marker median expressions.
}
\details{
For each marker, median expressions are discretized in severals categories corresponding to the heat intensities. 
This number of categories is provided using 'num' parameter.

By default the 'pheno.table' parameter is null and will be automatically compute using the 'computePhenoTable' function.
}

