% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{phenoViewer}
\alias{phenoViewer}
\title{Visualization of cluster phenotypes}
\usage{
phenoViewer(Results, samples = NULL, clusters = NULL, markers = NULL,
  show.mean = "both", show.on_device = TRUE, sort.markers = TRUE)
}
\arguments{
\item{Results}{a Results object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized (by default all clusters will be displayed)}

\item{markers}{a character vector specifying the markers to be displayed assignments}

\item{show.mean}{a character specifying if marker means expression should be displayed, possible value are among: "none", "only" or "both"}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}

\item{sort.markers}{a logical specifying if the markers must be sorted by names in the representation}
}
\value{
a 'ggplot' object
}
\description{
The Pheno Viewer aims to visualize the phenotype of single cluster or a set of combined clusters.
This representation displays median marker expressions of each sample using parallel coordinates. 
In such viewer, each line corresponds to a biological sample and lines are positioned on a space where the X-axis represents the cell markers and the Y-axis represents the marker expressions. 
If biological conditions have been assigned to the biological samples, then samples belonging to the same condition will be shown using the same color.

Marker expressions from several clusters can be combined by providing one or several cluster names to the `clusters` parameter.
In this case, the displayed marker expressions are the means of the median expressions for each cluster of each marker.

Markers used as clustering markers are shown in blue.
A dashed line indicates the mean marker expressions of all samples.
Importantly, a grey ribbon indicates the bounds of marker expressions in the overall dataset.
The visualization can be restricted to specific markers using the `markers` parameter.
}
\details{
The ranges of value between marker bounds (using the 'bounds' slot) will be displayed using a grey ribbon.

The 'show.mean' parameter allows to visualize three kinds of information:
\itemize{
\item "none" value will show marker median expressions for each selected samples;
\item "only" value will show only the mean of median maker expressions for all selected samples (displayed as black dashed line);
\item "both" value will show marker median expressions for each selected samples together with the mean of median maker expressions for all selected samples.
}
}

