% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,AC,ANY-method}
\alias{plot,CC,ANY-method}
\alias{plot,DEC,ANY-method}
\alias{plot,EnrichmentProfiles,ANY-method}
\alias{plot,PhenoProfiles,ANY-method}
\title{Graphical representation for some SPADEVizR objects}
\usage{
\S4method{plot}{DEC,ANY}(x, y, ...)

\S4method{plot}{AC,ANY}(x, y, ...)

\S4method{plot}{CC,ANY}(x, y, ...)

\S4method{plot}{PhenoProfiles,ANY}(x, y, ...)

\S4method{plot}{EnrichmentProfiles,ANY}(x, y, ...)
}
\arguments{
\item{x}{a 'AC', 'DEC', 'CC', 'PhenoProfiles' or 'EnrichmentProfiles' object}
}
\value{
a 'ggplot' object
}
\description{
This function generates a graphical representation for 'AC', 'DEC', 'CC', 'PhenoProfiles' and 'EnrichmentProfiles' objects.
}

