% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,AC,missing-method}
\alias{plot,CC,missing-method}
\alias{plot,DEC,missing-method}
\alias{plot,EnrichmentProfiles,missing-method}
\alias{plot,PhenoProfiles,missing-method}
\title{Graphical representation for some SPADEVizR objects}
\usage{
plot(x, y = NULL, ...)

\S4method{plot}{DEC,missing}(x, y = NULL, ...)

\S4method{plot}{AC,missing}(x, y = NULL, ...)

\S4method{plot}{CC,missing}(x, y = NULL, ...)

\S4method{plot}{PhenoProfiles,missing}(x, y = NULL, ...)

\S4method{plot}{EnrichmentProfiles,missing}(x, y = NULL, ...)
}
\arguments{
\item{x}{a 'AC', 'DEC', 'CC', 'PhenoProfiles' or 'EnrichmentProfiles' object}
}
\value{
a 'ggplot' object
}
\description{
This function generates a graphical representation for 'AC', 'DEC', 'CC', 'PhenoProfiles' and 'EnrichmentProfiles' objects.
}

