% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,AC,missing-method}
\alias{plot,CC,missing-method}
\alias{plot,CCR,missing-method}
\alias{plot,DAC,missing-method}
\title{Graphical representation for some SPADEVizR objects}
\usage{
plot(x, y = NULL, ...)

\S4method{plot}{DAC,missing}(x, y = NULL, ...)

\S4method{plot}{AC,missing}(x, y = NULL, ...)

\S4method{plot}{CC,missing}(x, y = NULL, ...)

\S4method{plot}{CCR,missing}(x, y = NULL, ...)
}
\arguments{
\item{x}{a 'AC', 'DAC', 'CC' and 'CCR' object}

\item{y}{a supplementary parameter transmited respectively to 'abundantClustersViewer()', 'volcanoViewer()' or 'correlatedClustersViewer()' functions}

\item{...}{some supplementaries parameters transmited respectively to \code{\link[SPADEVizR]{abundantClustersViewer}}, \code{\link[SPADEVizR]{volcanoViewer}} or \code{\link[SPADEVizR]{correlatedClustersViewer}} functions}
}
\value{
a 'ggplot' object
}
\description{
This function generates a graphical representation for 'AC', 'DAC', 'CC', 'CCR' and objects.
}

