% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,Results,missing-method}
\alias{plot,AC,missing-method}
\alias{plot,DAC,missing-method}
\alias{plot,CC,missing-method}
\alias{plot,AP,missing-method}
\title{Visualization for all SPADEVizR objects}
\usage{
plot(x, y = NULL, ...)

\S4method{plot}{Results,missing}(x, y = NULL, ...)

\S4method{plot}{AC,missing}(x, y = NULL, ...)

\S4method{plot}{DAC,missing}(x, y = NULL, ...)

\S4method{plot}{CC,missing}(x, y = NULL, ...)

\S4method{plot}{AP,missing}(x, y = NULL, ...)
}
\arguments{
\item{x}{a 'Results', 'AC', 'DAC', 'CC' and 'AP' object}

\item{y}{unused parameter}

\item{...}{supplementary parameters transmitted to the 'heatmapViewer()', 'abundantClustersViewer()', 'volcanoViewer()', 'correlatedClustersViewer()' or 'circlesPackingViewer()' functions}
}
\value{
a 'ggplot' object
}
\description{
Generates a graphical representation for all SPADEVizR objects ('Results', 'AC', 'DAC', 'CC', and 'AP' objects).
}
\details{
'Results' objects are represented using the 'heatmapViewer()' function.
'AC' objects are represented using the 'abundantClustersViewer()' function. 
'DAC' objects are represented using the 'volcanoViewer()' function.
'CC' objects are represented using the 'correlatedClustersViewer()' function.
'AP' objects are represented using the 'circlesPackingViewer()' function.
}
