% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,AC-method}
\alias{print,AP-method}
\alias{print,CC-method}
\alias{print,DAC-method}
\alias{print,Results-method}
\title{Textual previews for all SPADEVizR objects}
\usage{
\S4method{print}{Results}(x)

\S4method{print}{AC}(x)

\S4method{print}{DAC}(x)

\S4method{print}{CC}(x)

\S4method{print}{AP}(x)
}
\arguments{
\item{x}{a SPADEVizR object}
}
\value{
none
}
\description{
Prints a preview for a SPADEVizR object.
}

