% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{profilesViewer}
\alias{profilesViewer}
\title{profilesViewer}
\usage{
profilesViewer(profile.object)
}
\arguments{
\item{profile.object}{a PhenoProfiles object or an EnrichmentProfiles object}
}
\value{
a 'ggplot' object
}
\description{
Generate a graph representation of PhenoProfiles classes
}
\details{
xxx
}

