% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{Results-class}
\alias{Results}
\alias{Results-class}
\title{Results class definition}
\description{
The Results object is a S4 object containing results of automatic gating. This object store mainly the count matrix and the cluster phenotypes. It is to note that Results is a super classe of the SPADEResult
}
\details{
The Results object is the core (super classe) of SPADEResults object.

This object is allows to store automatic gating results from other algorithms. 
The importX() function return a Result Object.

The cells.count dataframe have in the first column the cluster names or numeric ID and some columns with number of cells for each sample (with the sample names in colnames)
The marker.expressions dataframe have .. to continue
}
\section{Slots}{

\describe{
\item{\code{cells.count}}{a dataframe containing the number of cells for each cluster of each sample}

\item{\code{marker.expressions}}{a numerical dataframe containing marker median expressions for each cluster of each sample}

\item{\code{sample.names}}{a character vector containing the sample names}

\item{\code{marker.names}}{a character vector containing the markers names}

\item{\code{cluster.number}}{a numeric specifying the number of cell clusters}
}}

