% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Results-class}
\alias{Results-class}
\alias{Results}
\title{Results class definition}
\description{
The Results object is a S4 object containing cell clustering results. 

This object mainly stores the cluster abundance matrix (i.e. the number of cells associated to each sample for each cluster) and the cluster phenotypes matrix (i.e. the median expressions for each marker of each cluster).
 
In addition, this object can contain information about clustering results, such a SPADE tree.
}
\details{
The 'cluster.abundances' dataframe contains the number of cells associated to each sample for each cluster.
This dataframe stores the clusters in rows and the samples in columns.

The 'cluster.phenotypes' dataframe stores the median expressions for each marker of each cluster.
This dataframe stores in the first column the sample names, in the second column the cluster names, and in the others columns the maker median expressions.

The 'bounds' dataframe contains the marker expressions boundaries (minimum and maximum, or specific percentiles) for each marker.

The 'print()' and 'show()' can be used to display a summary of this object.
}
\section{Slots}{

\describe{
\item{\code{cluster.abundances}}{a dataframe containing the number of cells associated to each sample for each cluster}

\item{\code{cluster.phenotypes}}{a dataframe containing the median expressions for each marker of each cluster}

\item{\code{sample.names}}{a character vector containing the sample names}

\item{\code{cluster.names}}{a character vector containing the cluster names}

\item{\code{cluster.number}}{a numeric specifying the number of clusters}

\item{\code{marker.names}}{a character vector containing the marker names}

\item{\code{clustering.markers}}{a character vector specifying the markers that have been used by the clustering algorithms}

\item{\code{bounds}}{a numeric data.frame containing the marker expressions boundaries for each marker}

\item{\code{trans}}{a character specifying what transformation ("arcsinh", "logicle", or "none") will be applied on the cluster expression matrix ("arcsinh" by default)}

\item{\code{flowset}}{a flowSet object containing the imported SPADE FCS files}

\item{\code{fcs.files}}{a character vector containing the location of the imported FCS files}

\item{\code{graph}}{a igraph object containing the SPADE tree structure}

\item{\code{graph.layout}}{a numeric matrix containing the SPADE tree layout}

\item{\code{assignments}}{a dataframe containing annotations for each sample samples such as a biological condition ("bc"), a timepoint condition ("tp") or an individual ("ind") assignment}

\item{\code{th.min_cells}}{a numeric specifying the minimal number of cells that a cluster for a given samples needs to have to be taken into consideration in its phenotypical characterization}
}}

