% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,AC-method}
\alias{show,CC-method}
\alias{show,CCR-method}
\alias{show,DAC-method}
\alias{show,Results-method}
\alias{show,SPADEResults-method}
\title{Textual previews for SPADEVizR objects}
\usage{
\S4method{show}{Results}(object)

\S4method{show}{SPADEResults}(object)

\S4method{show}{AC}(object)

\S4method{show}{DAC}(object)

\S4method{show}{CC}(object)

\S4method{show}{CCR}(object)
}
\arguments{
\item{object}{a SPADEVizR object}
}
\value{
none
}
\description{
Show a previews for a SPADEVizR object.
}

