% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_objects.R
\docType{class}
\name{SPADEResults-class}
\alias{SPADEResults}
\alias{SPADEResults-class}
\title{SPADEResults class definition}
\description{
The 'SPADEResults' object is a S4 object containing cell clustering results obtained from SPADE.


This object inherits from the 'Result' object and stores the count matrix (i.e. the number of cells associated with each cluster of each sample) and the cell cluster phenotypes (i.e. the marker median expressions for each cluster). 
In addition to the 'Result' object, the 'SPADEResults' object contains information about SPADE clustering results, such as the SPADE tree, the clustering makers and the FCS files.
}
\details{
The 'print()' and 'show()' can be used to display a summury of this object. Moreover all information about this object could be saved as a tab separated file using the 'export()' method.
This object is returned by the 'importSPADEResult()' function.
}
\section{Slots}{

\describe{
\item{\code{cells.count}}{a dataframe containing the number of cells for each cluster of each sample}

\item{\code{marker.expressions}}{a numerical dataframe containing marker median expressions for each cluster of each sample}

\item{\code{sample.names}}{a character vector containing the sample names}

\item{\code{marker.names}}{a character vector containing the markers names}

\item{\code{cluster.number}}{a numeric specifying the number of cell clusters}

\item{\code{use.raw.medians}}{a logical specifying if the marker expressions correspond to the raw or transformed data}

\item{\code{dictionary}}{a two column data.frame providing the correspondence between the original marker names (first column) and the real marker names (second column)}

\item{\code{marker.clustering}}{a logical vector specifying marker that have been used during the clustering precedure}

\item{\code{flowset}}{a flowSet object containing the imported SPADE FCS file}

\item{\code{fcs.files}}{a character vector containing the absolute path of the original FCS files}

\item{\code{quantiles}}{a numeric data.frame containing the quantiles for each each markers of cluster}

\item{\code{graph}}{a igraph object containing the SPADE tree}

\item{\code{graph.layout}}{a numeric matrix containing the layout of the SPADE tree}
}}

