% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{streamgraphViewer}
\alias{streamgraphViewer}
\title{streamgraphViewer}
\usage{
streamgraphViewer(Results, order = NULL, clusters = NULL)
}
\arguments{
\item{Results}{a SPADEResults or Results object}

\item{order}{a named vector a named vector providing the correspondence between a sample name (in rownames) and an integer ordering samples in numeric (NA to exclude this sample)}

\item{clusters}{a numerical vector containing the clusters to use in the representation, by default all clusters will be use}
}
\value{
a 'ggplot' object
}
\description{
Generate a streamgraph representation showing the evolution of cell count in clusters across samples
}
\details{
xxx
}

