% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{streamgraphViewer}
\alias{streamgraphViewer}
\title{Visualization of cluster abundance dynamics}
\usage{
streamgraphViewer(Results, samples = NULL, clusters = NULL,
  use.relative = FALSE, show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{clusters}{a character vector containing the clusters names to be visualized}

\item{use.relative}{a logical specifying if the visualization should be performed on relative abundance}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
The Streamgraph Viewer aims to visualize both absolute and relative abundance of clusters across the samples.
This representation displays cell abundances using a stacked area graph which is placed around a central axis.

The cell clusters to visualize must be specified using the `clusters` parameter. 
Moreover, samples to be represented and their orders can be specified using the `samples` parameter.
}
\details{
The order of samples in the 'samples' vector correspond to the order where the sample will be displayed.
}
