% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{treeViewer}
\alias{treeViewer}
\title{Visualization of combined SPADE trees}
\usage{
treeViewer(SPADEResults, samples = NULL, highlight = NULL, marker = NULL,
  show.on_device = TRUE)
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{highlight}{an AC, DAC or CC object to highlight identified significant clusters in the SPADE tree}

\item{marker}{a character specifying the marker name to display}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a list of 'ggplot' objects
}
\description{
Generates a tree representation showing combined SPADE trees.
}
\details{
The size of tree nodes are related to the number of cells in each cluster. 
If the 'stat.object' parameter is provided node outlines are colored according to clusters significance.
If the 'marker' parameter is provided, the nodes are colored according to mean expression for the selected marker using selected samples.
}

