% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_viewers.R
\name{treeViewer}
\alias{treeViewer}
\title{Visualization of combined SPADE trees}
\usage{
<<<<<<< HEAD
treeViewer(SPADEResults, samples = NULL, highlight = NULL, marker = NULL,
  show.on_device = TRUE)
=======
treeViewer(SPADEResults, samples = NULL, stat.object = NULL,
  marker = NULL)
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\arguments{
\item{SPADEResults}{a SPADEResults object (Results object is not accepted)}

\item{samples}{a named vector providing the correspondence between samples name (in rowname) and the logical value TRUE to use these samples (all samples by default)}

<<<<<<< HEAD
\item{highlight}{an AC, DEC or CC object to highligth identified significant clusters in the SPADE tree}

\item{marker}{a character specifying the marker name to display}

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
=======
\item{stat.object}{an AC, DEC or CC object to highligth identified significant clusters in the SPADE tree}

\item{marker}{a character specifying the marker name to display}
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\value{
a list of 'ggplot' objects
}
\description{
Generates a tree representation showing combined SPADE trees.
}
\details{
The size of tree nodes are related to the number of cells in each cluster. 
If the 'stat.object' parameter is provided node outlines are colored according to clusters signifiance.
If the 'marker' parameter is provided, the nodes are colored according to mean expression for the selected marker using selected samples.
}

