% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{treeViewer}
\alias{treeViewer}
\title{Visualization of combined SPADE trees}
\usage{
treeViewer(Results, samples = NULL, highlight = NULL, marker = NULL,
  show.on_device = TRUE)
}
\arguments{
\item{Results}{a 'Results' object (with 'graph' and 'graph.layout' slots not null)}

\item{samples}{a character vector providing the sample names to used (all samples by default)}

\item{highlight}{an AC, DAC or CC object to highlight identified significant clusters in the SPADE tree}

\item{marker}{a character specifying the marker to be displayed}

\item{show.on_device}{a logical specifying if the representation will be displayed on device}
}
\value{
a 'ggplot' object
}
\description{
The Tree Viewer aims to visualize the SPADE tree representations. 
This representation displays the SPADE cell clusters using this minimal spanning tree layout computed by SPADE. 
In such tree, each node represents a cell cluster and nodes are linked based on theirs phenotype similarities.
This viewer improves the original SPADE tree representations by allowing to combine SPADE trees from multiple samples. 

Significant clusters can be highlighted (node borders are then colored in blue) by providing a `AC`, `DAC`, or `CC` object (using the `highlight` parameter).
As with the original SPADE tree representations, nodes can be colored based on the marker median expression of a specific marker (using the `marker` parameter).
}
\details{
The size of tree nodes is proportional to the number of cells in each cluster. 
If the 'stat.object' parameter is provided node outlines are colored according to clusters significance.
}
