% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{volcanoViewer}
\alias{volcanoViewer}
\title{Volcano Plot Viewer}
\usage{
volcanoViewer(DEC = NULL, fc.log2 = TRUE, cluster.size = TRUE,
  all.label = FALSE)
}
\arguments{
\item{DEC}{object DEC provided by the function computeDEC}

\item{fc.log2}{a logicial specifing if foldchange or log2(foldchange) is use}

\item{cluster.size}{a logicial specifing if points size are related to cell count or not (TRUE by default)}

\item{all.label}{a logicial specifing if all cluster labels must be show or just significant cluster}

\item{SPADEResults}{the SPADEViewer result object}
}
\value{
a ggplot object
}
\description{
Generate a Volcano plot representation based on Differentially Enriched Clusters (DEC) of The SPADE result object.
}
\details{
xxx
}

