% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{volcanoViewer}
\alias{volcanoViewer}
\title{Visualization of differentially enriched clusters}
\usage{
volcanoViewer(DEC = NULL, fc.log2 = TRUE, show.cluster.sizes = TRUE,
<<<<<<< HEAD
  show.all_labels = FALSE, show.on_device = TRUE)
=======
  show.all_labels = FALSE)
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\arguments{
\item{DEC}{an object of class 'DEC' (object returned by the 'computeDEC()' function)}

\item{fc.log2}{a logical specifying if fold-change or log2(fold-change) is use}

\item{show.cluster.sizes}{a logical specifying if dot sizes are proportional to cell counts}

\item{show.all_labels}{a logical specifying if all cluster labels must be show or just significant cluster}
<<<<<<< HEAD

\item{show.on_device}{a logical specifying if the respresentation will be displayed on device}
=======
>>>>>>> 43ace80f59a8e1b7a65c6f88e2a134a1e91a2f23
}
\value{
a 'ggplot' object
}
\description{
Generates a Volcano plot representation showing for each cluster
}
\details{
By default, only significant differentially enriched clusters are labeled. Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}

