% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewers.R
\name{volcanoViewer}
\alias{volcanoViewer}
\title{Visualization of differentially abundant clusters}
\usage{
volcanoViewer(DAC = NULL, fc.log2 = TRUE, show.cluster_sizes = TRUE,
  show.all_labels = FALSE, show.on_device = TRUE,
  max.dots_size = max(DAC@cluster.size), y.max = NULL)
}
\arguments{
\item{DAC}{an object of class 'DAC' (object returned by the 'computeDAC()' function)}

\item{fc.log2}{a logical specifying if the fold-change or the log2(fold-change) must be used}

\item{show.cluster_sizes}{a logical specifying if dot sizes are proportional to number of associated cells}

\item{show.all_labels}{a logical specifying if all cluster labels must be shown. Only labels of significant clusters are displayed otherwise}

\item{show.on_device}{a logical specifying if the ggplot representation must be displayed on the device}

\item{max.dots_size}{a numeric specifying the number of associated cells in the largest dot}

\item{y.max}{a numeric specifying the maximal value of the y axis}
}
\value{
a 'ggplot' object
}
\description{
Generates a volcano plot representation showing for each cluster: its mean abundance fold-change and associated p-value. 

This representation displays the p-value (shown as -log10(p-value)) in the Y-axis and the fold-change of cell abundances, in the X-axis in a two-dimensional chart. 
Each dot in the representation corresponds to a cell cluster, and both p-value and fold-change thresholds are shown using red dashed lines.
Differentially Abundant Clusters are highlighted in red and labeled.
The size of dots is proportional to the total number of associated cells in the 2 conditions merged.
}
\details{
By default, only significant differentially abundant clusters are labeled. 
Labels for all clusters can be displayed by setting the 'all.label' parameter to TRUE.
}

